% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Add column \code{id} to a dataframe by pasting values from other columns.}
\usage{
add_id(.data, vars, name = "id", sep = "_")
}
\arguments{
\item{.data}{A dataframe.}

\item{vars}{Vector of names of columns to paste for creating the new id
column.}

\item{name}{String giving the name of the new id column.}

\item{sep}{String to separate the pasted values of \code{vars}.}
}
\value{
A dataframe with an additional column.
}
\description{
Add column \code{id} to a dataframe by pasting values from other columns.
}
\examples{
.data <- tibble::tibble(x = 1:2, y = c("a", "b"), z = c(TRUE, FALSE))
add_id(.data, c("x", "y", "z"))
add_id(.data, c("x", "z"))
add_id(.data, c("x", "z"), sep = 1)

}
\seealso{
Other functions to add columns to dataframes: \code{\link{add_status_tree}},
  \code{\link{add_subquad}}, \code{\link{add_var}}
}
\concept{functions to add columns to dataframes}
