% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{add_var}
\alias{add_var}
\alias{add_lxly}
\alias{add_qxqy}
\alias{add_index}
\alias{add_col_row}
\alias{add_hectindex}
\alias{add_quad}
\alias{add_col_row2}
\title{Add columns \code{lx/ly}; \code{QX/QY}; \code{index}; \code{col/row}; and \code{hectindex}, \code{quad}.}
\usage{
add_lxly(x, gridsize = 20, plotdim = NULL)

add_qxqy(x, gridsize = 20, plotdim = NULL)

add_index(x, gridsize = 20, plotdim = NULL)

add_col_row(x, gridsize = 20, plotdim = NULL)

add_hectindex(x, gridsize = 20, plotdim = NULL)

add_quad(x, gridsize = 20, plotdim = NULL, start = 1, width = 2)

add_col_row2(x)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{gridsize}{The gridsize of the census plot (commonly 20 m).}

\item{plotdim}{The global dimensions of the census plot (i.e. the
maximum possible values of \code{gx} and \code{gy}).}

\item{start}{\code{1} or \code{0}, indicating how to label the first plot-column.}

\item{width}{Number; width to pad the labels of plot-columns and -rows
(passed to \code{\link[stringr:str_pad]{stringr::str_pad()}}).}
}
\value{
A modified version of the dataframe \code{x} with the additional
variable(s) \code{var}.
}
\description{
These functions add columns to position trees/stems in a forest plot. They
work with ViewFull-tables and census-tables (tree and stem).
}
\details{
These functions mostly wrap legacy code from the \href{http://ctfs.si.edu/Public/CTFSRPackage/}{CTFS R Package}.
}
\examples{
x <- tibble::tibble(gx = c(0, 50, 999.9, 1000), gy = gx/2)
add_lxly(x)
add_qxqy(x)
add_index(x)
add_hectindex(x)
add_quad(x)
add_quad(x, start = 0)

# `width` gives the nuber of digits to pad the label of plot-rows and
# plot-columns, e.g. 3 pads plot-rows with three zeros and plot-columns with
# an extra trhree zeros, resulting in a total of 6 zeros.
add_quad(x, start = 0, width = 3)
add_col_row(x)

# Column and row from QuadratName
x <- tibble::tribble(
  ~QuadratName,
  "0001",
  "0011",
  "0101",
  "1001"
)
add_col_row2(x)
}
\seealso{
\code{\link[stringr:str_pad]{stringr::str_pad()}}.

Other functions to add columns to dataframes.: \code{\link{add_status_tree}},
  \code{\link{add_subquad}}

Other functions for ForestGEO data.: \code{\link{add_status_tree}},
  \code{\link{add_subquad}}

Other functions for fgeo census.: \code{\link{add_status_tree}},
  \code{\link{guess_plotdim}}, \code{\link{pick_drop}}

Other functions for fgeo vft.: \code{\link{add_status_tree}},
  \code{\link{guess_plotdim}}, \code{\link{pick_drop}},
  \code{\link{pick_plotname}}
}
\concept{functions for ForestGEO data.}
\concept{functions for fgeo census.}
\concept{functions for fgeo vft.}
\concept{functions to add columns to dataframes.}
