% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuses.R
\name{as_censuses}
\alias{as_censuses}
\alias{as_censuses.default}
\alias{as_censuses.list}
\alias{as_censuses.tbl_df}
\title{Create objects of class \code{censuses_*}, where \code{*} depends on the input.}
\usage{
as_censuses(.data)

\method{as_censuses}{default}(.data)

\method{as_censuses}{list}(.data)

\method{as_censuses}{tbl_df}(.data)
}
\arguments{
\item{.data}{A ForestGEO-like dataset.}
}
\value{
An object of class \code{census_*}, where \code{*} depends on the input.
}
\description{
Create objects of class \code{censuses_*}, where \code{*} depends on the input.
}
\examples{
censuses_lst <- as_censuses(list(
  c1 = tibble(dbh = 1:2),
  c2 = tibble(dbh = 8:9)
))
class(censuses_lst)
}
\seealso{
Other functions to construct fgeo classes: \code{\link{fgeo_elevation}},
  \code{\link{fgeo_topography}}
}
\concept{functions to construct fgeo classes}
