% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_is_installed.R
\name{assert_is_installed}
\alias{assert_is_installed}
\title{Assert a package is installed.}
\usage{
assert_is_installed(pkg)
}
\arguments{
\item{pkg}{Character vector giving the name of a package.}
}
\value{
An error if \code{pkg} is not installed or invisible \code{pkg} if it is.
}
\description{
Assert a package is installed.
}
\examples{
assert_is_installed("base")
try(assert_is_installed("bad"))
}
\keyword{internal}
