% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_treeid_pick_dbh.R
\name{by_treeid_pick_dbh}
\alias{by_treeid_pick_dbh}
\alias{by_treeid_pick_dbh_min}
\alias{by_treeid_pick_dbh_max}
\title{Pick only the one stem per treeid (per censusid).}
\usage{
by_treeid_pick_dbh_min(.x)

by_treeid_pick_dbh_max(.x)
}
\arguments{
\item{.x}{A dataframe; particularly a ForestGEO census or ViewFullTable.}
}
\value{
A dataframe with one row per per treeid per censusid and a single
plotname.
}
\description{
\code{by_treeid_pick_dbh_max()} and \code{by_treeid_pick_dbh_min()} pick the stem with the
maximum and minimum dbh per treeid per censusid. It intentionally errs if
it detects multiple distinct values of plotname (i.e. if the data is a
ViewFullTable).
}
\examples{
census <- tibble::tribble(
  ~dbh,   ~sp, ~treeID, ~stemID,
    10, "sp1",     "1",   "1.1",
   100, "sp1",     "1",   "1.2",
    22, "sp2",     "2",   "2.1",
    99, "sp2",     "2",   "2.2",
    99, "sp2",     "2",   "2.3",
    NA, "sp2",     "2",   "2.4"
)

by_treeid_pick_dbh_max(census)

by_treeid_pick_dbh_min(census)
}
\seealso{
Other functions to pick or drop rows of a dataframe.: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}}, \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{functions to pick or drop rows of a dataframe.}
