% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allquadratslopes.R
\name{calcslope}
\alias{calcslope}
\title{Internal.}
\usage{
calcslope(z, gridsize = 20)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{allquadratslopes}},
  \code{\link{findborderquads}}, \code{\link{quadslope}}
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
