% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_structure_elev.R
\name{clean_structure_elev}
\alias{clean_structure_elev}
\title{If necessary, pull elevation data and rename \code{x} and \code{y} to \code{gx} and \code{gy}.}
\usage{
clean_structure_elev(elevation)
}
\arguments{
\item{elevation}{A dataframe or list of ForestGEO's elevation data.}
}
\value{
A dataframe
}
\description{
This functions helps standarize elevation data. Elevation data may be stored
as a dataframe in the element \code{col} of a list or may be directly the
dataframe. Also, the names may be \code{x} and \code{y} wich is inconsistent with the
names \code{gx} and \code{gy} of census datasets. Whatever the structure of the input
(either a dataframe or a list) and whatever the names (\code{x} and \code{y}, or \code{gx}
and \code{gy}), this function outputs a dataframe with names \code{gx} and \code{gy}, and
checks that the variable \code{elev} is present.
}
\examples{
# Dataframe input
elev_df <- bciex::bci_elevation
str(elev_df)

elev_clean <- clean_structure_elev(elev_df)
str(elev_clean)

# List input
elev_list <- list(col = elev_df, other = "stuff")
str(elev_list)

elev_clean <- clean_structure_elev(elev_list)
str(elev_clean) 
}
