% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{\itemize{
\item \code{cluster_elevation()} outputs one additional column, \code{cluster}, calculated
by hierarchical clustering of the topographic metrics calculated by
\code{fgeo_topography()}. \code{cluster_elevation()} first calculates a
dissimilarities object (with \code{\link[stats:dist]{stats::dist()}} and all its defaults), then it
calculates a tree (with \code{\link[stats:hclust]{stats::hclust()}} and all its defaults), and finally
cuts the tree in \code{n} groups (with \code{\link[stats:cutree]{stats::cutree()}}).
}}
\usage{
cluster(.data, ...)
}
\arguments{
\item{.data}{A dataframe.}

\item{...}{Other arguments passed to methods.}
}
\description{
\itemize{
\item \code{cluster_elevation()} outputs one additional column, \code{cluster}, calculated
by hierarchical clustering of the topographic metrics calculated by
\code{fgeo_topography()}. \code{cluster_elevation()} first calculates a
dissimilarities object (with \code{\link[stats:dist]{stats::dist()}} and all its defaults), then it
calculates a tree (with \code{\link[stats:hclust]{stats::hclust()}} and all its defaults), and finally
cuts the tree in \code{n} groups (with \code{\link[stats:cutree]{stats::cutree()}}).
}
}
\examples{
elev_ls <- fgeo.data::luquillo_elevation
topo <- fgeo_topography(elev_ls, gridsize = 20)
cluster(topo, n = 4)
}
