% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_censusid.R
\name{collapse_censusid}
\alias{collapse_censusid}
\title{Collapse the values of CensusID.}
\usage{
collapse_censusid(x)
}
\arguments{
\item{x}{A dataframe -- specifically a ViewFullTable.}
}
\value{
A modified version of \code{x}, most likely with less rows.
}
\description{
This function is useful to summarize the status history of a tree tag across
censuses. It is particularly useful to avoid duplicated tree tags on maps.
}
\examples{
library(dplyr)

x <- dplyr::tribble(
  ~CensusID, ~Tag, ~Status,
        "1",  "1", "alive",
        "1",  "1",  "dead",
        "1",  "2",  "dead",
        "1",  "2",  "dead",
        "2",  "1", "alive",
        "2",  "1", "alive",
        "2",  "2", "alive",
        "2",  "2",  "dead"
)

collapse_censusid(x)

with_status_tree <- fgeo.tool::add_status_tree(x, "alive", "dead")
collapsed_cns <- collapse_censusid(with_status_tree)
unique(select(collapsed_cns, -Status))
}
