% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{conv_unit_at}
\alias{conv_unit_at}
\title{Convert units of a vector or columns of a dataframe.}
\usage{
conv_unit_at(x, .at, from, to)
}
\arguments{
\item{x}{A vector or dataframe.}

\item{.at}{A character vector of names or a numeric vector of
positions. Only those elements corresponding to \code{.at} will be
modified.}

\item{from}{the unit in which the measurement was made.}

\item{to}{the unit to which the measurement is to be converted.}
}
\value{
An object with the same structure as \code{x} (vector or dataframe).
}
\description{
\code{conv_unit()} of a vector or \code{conv_unit_at()} specific columns of a
dataframe.
}
\examples{
# For all units see ?measurements::conv_unit()
conv_unit(10, "mm2", "hectare")
conv_unit(1:3, from = "m", to = "mm")
conv_unit(1:3, "mph", "kph")

dfm <- data.frame(dbh = c(10, 100), name = c(10, 100))
conv_unit_at(dfm, .at = "dbh", from = "mm2", to = "hectare")
# All columns
conv_unit_at(dfm, .at = c("dbh", "name"), from = "mm2", to = "hectare")
# Same
conv_unit_at(dfm, .at = names(dfm), from = "mm2", to = "hectare")
}
\seealso{
\code{\link[measurements:conv_unit]{measurements::conv_unit()}}, \code{\link[purrr:map_at]{purrr::map_at()}}.
}
