% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\alias{convert_units_at}
\title{Convert units of a vector or columns of a dataframe.}
\usage{
convert_units(x, from, to)

convert_units_at(x, .at, from, to)
}
\arguments{
\item{x}{A vector or dataframe.}

\item{from, to}{String of the units to convert from/to as in \link[units:ud_units]{units::ud_units}.}

\item{.at}{A character vector of names or a numeric vector of
positions. Only those elements corresponding to \code{.at} will be
modified.}
}
\value{
An object with the same structure as \code{x} (vector or dataframe).
}
\description{
\code{convert_units()} of a vector or \code{convert_units_at()} specific columns of a
dataframe.
}
\examples{
# View valid units with: View(units::valid_udunits())
convert_units(1:3, from = "mm", to = "m")
convert_units(1:3, "hectare", "m^2")
convert_units(1:3, "km/h", "m/h")

dfm <- data.frame(a = 1:2, b = 2:3)
convert_units_at(dfm, from = "m", to = "mm", .at = c("a", "b"))
convert_units_at(dfm, from = "m", to = "mm", .at = names(dfm))
convert_units_at(dfm, from = "m", to = "mm", .at = 2)
}
\seealso{
\code{\link[units:valid_udunits]{units::valid_udunits()}}, \code{\link[purrr:map_at]{purrr::map_at()}}.
}
