% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_habitat.R
\name{create_habitat}
\alias{create_habitat}
\title{Create habitat data from elevation data.}
\usage{
create_habitat(elevation, gridsize, n)
}
\arguments{
\item{elevation}{List containing elevation data.}

\item{gridsize}{Number to round \code{x} and \code{y} by. Commonly, \code{gridsize = 20}.}

\item{n}{Number of elevation groups (habitats) to cut elevation data by.
Commonly, \code{n = 4}.}
}
\value{
A dataframe of habitat data.
}
\description{
Create habitat data from elevation data.
}
\examples{
create_habitat(fgeo.data::luquillo_elevation, gridsize = 20, n = 4)
}
