% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listed_df.R
\name{csv_listed_df}
\alias{csv_listed_df}
\title{Save each dataframe in a list to a different .csv file.}
\source{
Adappted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
csv_listed_df(df_list, dir, prefix = NULL)
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{dir}{Character; the directory where the files will be saved.}

\item{prefix}{Character; a prefix to add to the file names.}
}
\description{
A useful complement of this function is \code{\link[=xl_list_sheets]{xl_list_sheets()}}.
}
\examples{
path_to_excel_workbook <- system.file(
  "extdata", "example.xlsx", package = "fgeo.tool"
)

df_list <- xl_list_sheets(path_to_excel_workbook)
str(df_list)

output <- tempdir()
csv_listed_df(df_list, output, prefix = "myfile-")

files <- dir(output)
files[grepl(".csv$", files)]
}
\seealso{
xl_list_sheets.

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{xl_list_sheets}}
}
