% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicated_var.R
\name{detect_duplicated_var}
\alias{detect_duplicated_var}
\title{Factory of functions to detect duplicated values of a variable by groups.}
\usage{
detect_duplicated_var(var)
}
\arguments{
\item{var}{Bare name of a variable.}
}
\value{
A closure with the formal argument \code{.data} that returns \code{TRUE} or
when duplicated values of \code{var} are detected or \code{FALSE} if not.
}
\description{
Factory of functions to detect duplicated values of a variable by groups.
}
\examples{
detect_duplicated_treeid_by_group <- detect_duplicated_var(treeid)
tree <- tibble::tibble(treeID = c(1, 1))
detect_duplicated_var(treeid)(tree)
}
\seealso{
Other functions for developers.: \code{\link{find_data_of_class}},
  \code{\link{flag_duplicated_var}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}},
  \code{\link{type_ensure}}

Other function factories.: \code{\link{flag_duplicated_var}}
}
\concept{function factories.}
\concept{functions for developers.}
\concept{predicates.}
