% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_output.R
\name{dfs_to_csv}
\alias{dfs_to_csv}
\title{Save each dataframe in a list to a different .csv file.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
dfs_to_csv(df_list, dir, prefix = NULL)
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{dir}{Character; the directory where the files will be saved.}

\item{prefix}{Character; a prefix to add to the file names.}
}
\description{
A useful complement of this function is \code{\link[=xlsheets_to_dfs]{xlsheets_to_dfs()}}.
}
\examples{
dfs <- xlsheets_to_dfs(tool_example("multiple_sheets.xlsx"))

# Saving the output to a temporary file
output <- tempdir()
dfs_to_csv(dfs, output, prefix = "myfile-")

# Look inside the output directory to confirm it worked
dir(output, pattern = "myfile")
}
\seealso{
xlsheets_to_dfs

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{xlsheets_to_dfs}}
}
\concept{functions to handle multiple spreadsheets of an excel workbook.}
