% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{dir_list}
\alias{dir_list}
\alias{rdata_list}
\alias{rds_list}
\alias{csv_list}
\alias{delim_list}
\alias{tsv_list}
\alias{xl_list}
\alias{xlbooks_list}
\title{Import multiple files (.csv, excel, .Rdata, ...) from a directory into a list.}
\usage{
rdata_list(path_dir, ...)

rds_list(path_dir, ...)

csv_list(path_dir, ...)

delim_list(path_dir, ...)

tsv_list(path_dir, ...)

xl_list(path_dir, ...)

xlbooks_list(path_dir, ...)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function.}
}
\value{
A list of dataframes.
}
\description{
These functions read from a specific directory where every file has its own extension
as indicated by each function's name. Notice that function names have the format
input_output, i.e. file-extension_list. If none of these functions do what
you want, create your own with \code{\link[=read_with]{read_with()}}.
}
\examples{
path_rdata <- tool_example("rdata")
path_rdata
dir(path_rdata)
rdata_list(path_rdata)

path_rds <- tool_example("rds")
path_rds
dir(path_rds)
rds_list(path_rds)

path_csv <- tool_example("csv")
path_csv
dir(path_csv)
csv_list(path_csv)

path_tsv <- tool_example("tsv")
path_tsv
dir(path_tsv)
tsv_list(path_tsv)

# Weird: Tab separated columns in a file with .csv extension
path_weird <- tool_example("weird")
dir(path_weird)
# Extension is .csv, but this is not what you want
csv_list(path_weird)
# Use this instead
delim_list(path_weird, delim = "\\t")

path_xl <- tool_example("xl")
path_xl
dir(path_xl)
xl_list(path_xl)

path_books <- tool_example("multiple_workbooks")
dir(path_books)
xlbooks_list(path_books)
}
\seealso{
Other general functions to import data: \code{\link{rdata_df}},
  \code{\link{read_with}}, \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
