% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_twice_dead.R
\name{drop_twice_dead}
\alias{drop_twice_dead}
\title{Remove trees found dead in both the last and previous last censuses.}
\usage{
drop_twice_dead(vft)
}
\arguments{
\item{vft}{A dataframe -- specifically a ForestGEO ViewFullTable with
variables \code{status_tree} (see \code{\link[=add_status_tree]{add_status_tree()}}).}
}
\value{
A modified version of the input data set:
* With the rows removed of all censuses except the last two.
* With the rows removed of trees found dead on both the last and previous
last censuses.
}
\description{
Removes trees that were found dead both in the last and previous last
censuses. Thus the resulting data set contains trees that in the last census
were found either alive or dead for the first time.
}
\examples{
vft <- tibble::tribble(
  ~CensusID, ~Treeid,   ~Status,
  1,    1,   "alive",  # Irrelevant: not one of the last two censuses
  1,    1,    "dead",  # 
  
  1,    2,   "alive",  # Irrelevant: not one of the last two censuses
  1,    2,   "alive",  # 
  
  2,    1,   "alive",  # Tree is alive: at lease one stem is alive
  2,    1,    "dead",  # 
  
  2,    2,    "dead",  # Tree is dead: all stems are dead. Notice that
  2,    2,    "dead",  # this tree is also dead in census 3.
  
  3,    1,   "alive",  # Tree is alive: at lease one stem is alive
  3,    1,    "dead",  # 
  
  3,    2,    "dead",  # Tree is dead: all stems are dead. Notice that
  3,    2,    "dead"   # this tree is also dead in census 2.
)

# `Status` refers to stems while `status_tree` refers to trees.
vft <- add_status_tree(vft, status_a = "alive", status_d = "dead")

# * Remove all censuses except the last two.
# * Remove trees found dead on both the last and previous last censuses.
drop_twice_dead(vft)
}
\seealso{
Other functions to pick or drop rows of a dataframe.: \code{\link{by_treeid_pick_dbh}},
  \code{\link{filter_status}}, \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{functions to pick or drop rows of a dataframe.}
