% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_path_dir.R
\name{example_path}
\alias{example_path}
\title{Path to directory containing example data.}
\usage{
example_path(path)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/.}
}
\value{
Path to directory containing example data.
}
\description{
Path to directory containing example data.
}
\examples{
example_path("two_files/new_stem_1.xlsx")
dirname(example_path("two_files/new_stem_1.xlsx"))
basename(example_path("two_files/new_stem_1.xlsx"))
}
