% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_habitat.R
\name{extract_from_habitat}
\alias{extract_from_habitat}
\alias{extract_gridsize}
\alias{extract_plotdim}
\title{Extract plot dimensions from habitat data.}
\usage{
extract_gridsize(habitats)

extract_plotdim(habitats)
}
\arguments{
\item{habitats}{Data frame giving the habitat designation for each 20x20
quadrat.}
}
\value{
\itemize{
\item \code{\link[=extract_plotdim]{extract_plotdim()}}: \code{plotdim} (vector of length 2);
\item \code{\link[=extract_gridsize]{extract_gridsize()}}: \code{gridsize} (scalar).
}
}
\description{
Extract plot dimensions from habitat data.
}
\examples{
library(fgeo.data)
extract_plotdim(luquillo_habitat)
extract_gridsize(luquillo_habitat)
}
\keyword{internal}
