% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.tool-package.R
\docType{package}
\name{fgeo.tool-package}
\alias{fgeo.tool}
\alias{fgeo.tool-package}
\title{fgeo.tool: Import and Manipulate 'ForestGEO' Data}
\description{
To help you access, transform, analyze, and visualize 'ForestGEO' data, we developed a collection of R packages (\url{https://forestgeo.github.io/fgeo/}). This package, in particular, helps you to easily import, filter, and modify 'ForestGEO' data. To learn more about 'ForestGEO' visit \url{https://forestgeo.si.edu/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forestgeo.github.io/fgeo.tool/}
  \item \url{https://github.com/forestgeo/fgeo.tool}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.tool/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Authors:
\itemize{
  \item Richard Condit \email{richardcondit@gmail.com}
  \item Suzanne Lao \email{laoz@si.edu}
  \item Anudeep Singh \email{anudeep7@gmail.com}
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
