% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.utils-package.R
\docType{package}
\name{fgeo.utils-package}
\alias{fgeo.utils}
\alias{fgeo.utils-package}
\title{fgeo.utils: Utility Tools for ForestGEO Packages}
\description{
Utility tools for forestgeo packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.utils}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.utils/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
