% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_elev.R
\name{fgeo_elev}
\alias{fgeo_elev}
\title{Construct elevation data}
\usage{
fgeo_elev(x)
}
\arguments{
\item{x}{Either a dataframe or a dataframe strored in the element \code{col} of a
list.}
}
\value{
A dataframe with names \code{x/gx}, \code{y/gy} and \code{elev}.
}
\description{
This function constructs an object of class fgeo_elev. It standarizes the
structure of elevation data to always output a dataframe with names \code{gx},
\code{gy} and \code{elev} -- or it dies trying.
}
\details{
The input may be a dataframe or a dataframe stored in the element \code{col} of a
list; the column names of the dataframe may be \code{gx} and \code{gy} or \code{x} and \code{y}.
}
\section{Acknowledgments}{

This function was inspired by David Kenfack.
}

\examples{
# Dataframe input
elev_df <- bciex::bci_elevation
str(elev_df)

elev <- fgeo_elev(elev_df)
str(elev)

# List input
elev_list <- list(col = elev_df, other = "stuff")
str(elev_list)

elev <- fgeo_elev(elev_list)
str(elev) 
}
