% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_elevation.R
\name{fgeo_elevation}
\alias{fgeo_elevation}
\title{Structure elevation data.}
\usage{
fgeo_elevation(x)
}
\arguments{
\item{x}{Either a dataframe or a dataframe stored in the element \code{col} of a
list.}
}
\value{
A dataframe with names \code{x/gx}, \code{y/gy} and \code{elev}.
}
\description{
This function constructs an object of class fgeo_elevation. It standardizes
the structure of elevation data to always output a dataframe with names \code{gx},
\code{gy} and \code{elev} -- or it dies trying.
}
\section{Input}{

The main input can be either the elevation list that ForestGEO delivers, or
the element \code{col} of such list -- which is a dataframe containing the
elevation data. Notice that the required arguments vary according to the main
input (the elevation list or the elevation dataframe). Whatever the input,
the dataframe containing the elevation data must have columns \code{gx} and \code{gy}
or \code{x} and \code{y}.
}

\section{Acknowledgments}{

This function was inspired by David Kenfack.
}

\examples{
# Input: Elevation dataframe
elevation_df <- fgeo.x::elevation$col
fgeo_elevation(elevation_df)

class(elevation_df)
class(fgeo_elevation(elevation_df))

names(elevation_df)
names(fgeo_elevation(elevation_df))

# Input: Elevation list
elevation_ls <- fgeo.x::elevation
fgeo_elevation(elevation_ls)

class(elevation_ls)
class(fgeo_elevation(elevation_ls))

names(elevation_ls)
names(fgeo_elevation(elevation_ls))

}
\concept{functions to construct fgeo classes}
\concept{habitat functions}
