% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldforms.R
\name{fieldforms_prepare}
\alias{fieldforms_prepare}
\title{Prepare data to create field forms.}
\usage{
fieldforms_prepare(x, y_q = 20, x_q = 20, x_sq = 5, y_sq = 5,
  subquad_offset = NULL)
}
\arguments{
\item{x}{A dataframe; particularly a ForestGEO ViewFullTable.}

\item{y_q}{Size in meters of a quadrat's side. For ForestGEO sites, a
common value is 20.}

\item{x_q}{Size in meters of a quadrat's side. For ForestGEO sites, a
common value is 20.}

\item{x_sq}{Size in meters of a subquadrat's side. For ForestGEO-CTFS
sites, a common value is 5.}

\item{y_sq}{Size in meters of a subquadrat's side. For ForestGEO-CTFS
sites, a common value is 5.}

\item{subquad_offset}{A number; either \code{-1} or \code{1}, to rest or add one unit
to the number of column of each subquadrat.}
}
\value{
A list of restructured dataframes.
}
\description{
This function creates a data structure suitable for \code{fieldforms_output()}. The data
should be from a single plot and of a single census.
}
\examples{
prepared <- fieldforms_prepare(top1quad)
}
