% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_to_df.R
\name{files_to_df}
\alias{files_to_df}
\alias{files_to_df}
\alias{csv_to_df}
\alias{files_to_df}
\alias{csv_to_df_lst}
\alias{files_to_df}
\alias{xl_to_df}
\alias{files_to_df}
\alias{xl_to_df_lst}
\title{Read all .csv or excel files from a directory into dataframes.}
\usage{
csv_to_df(input_dir)

csv_to_df_lst(input_dir)

xl_to_df(input_dir)

xl_to_df_lst(input_dir)
}
\arguments{
\item{input_dir}{String giving the directory containing the excel workbooks
to read from.}
}
\value{
The versions ending in _df output a single dataframe. The versions
ending in _lst output a list of dataframes.
}
\description{
Read all .csv, .xls, or .xlsx files from a directory into a dataframe or a
list of dataframes, where each element of the list is named as the source
file.
}
\section{Acknowledgment}{

Thanks to Jessica Shue for inspiring this functions.
}

\examples{
path <- system.file("extdata", "files/01.csv", package = "fgeo.tool")
input_dir <- fs::path_dir(path)
input_dir
dir(input_dir)

# Read all .csv files
csv_to_df(input_dir)
csv_to_df_lst(input_dir)

# Read all .xls or .xlsx files
xl_to_df_lst(input_dir)
xl_to_df(input_dir)
}
