% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Modify vectors, dataframes, or matrices to fill missing values.}
\usage{
fill_na(x, filler = 0)
}
\arguments{
\item{x}{A dataframe.}

\item{filler}{A filler; whatever you want to replace NA's with.}
}
\value{
A modified version of \code{x}.
}
\description{
This is a wrapper of \code{\link[base:replace]{base::replace()}} with a simplified interface and a name
that target the specific job of filling NA's. See also \code{tidyr::replace_na}
and \code{dplyr::coalesce}.
}
\examples{
x <- data.frame(x = c(NA, 1), y = c("a", NA), stringsAsFactors = FALSE)

a_dataframe <- x
fill_na(a_dataframe)
fill_na(a_dataframe, "")
fill_na(a_dataframe, "missing")

a_matrix <- as.matrix(x)
fill_na(a_matrix)

a_vector <- x$x
fill_na(a_vector)

a_list <- list(x, x, x)
lapply(a_list, fill_na)
}
\seealso{
Other general functions to edit data in place: \code{\link{lookup}},
  \code{\link{suffix_match}}, \code{\link{to_tidy_names}}
}
\concept{general functions to edit data in place}
