% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data_of_class.R
\name{find_data_of_class}
\alias{find_data_of_class}
\alias{show_data_of_class}
\alias{hide_data_of_class}
\title{Find and hide dataset names in a package that match a class.}
\usage{
show_data_of_class(package, .class)

hide_data_of_class(package, .class)
}
\arguments{
\item{package}{String; Package name to search datasets in.}

\item{.class}{Class of data to find or hide.}
}
\description{
Find and hide dataset names in a package that match a class.
}
\examples{
show_data_of_class("fgeo.tool", "tbl")
hide_data_of_class("fgeo.tool", "data.frame")
}
\seealso{
Other functions for developers.: \code{\link{detect_duplicated_var}},
  \code{\link{flag_duplicated_var}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}},
  \code{\link{type_ensure}}
}
\concept{functions for developers.}
