% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicated_var.R
\name{flag_duplicated_var}
\alias{flag_duplicated_var}
\title{Factory to flag duplicated values of a variable (message, warning or error).}
\usage{
flag_duplicated_var(.f, var)
}
\arguments{
\item{.f}{A condition function: One of message, warning, or stop.}

\item{var}{Bare name of a variable.}
}
\value{
A closure with the formal argument \code{.data} that returns the flag
produced by \code{.f} when duplicated values of \code{var} are detected.
}
\description{
Factory to flag duplicated values of a variable (message, warning or error).
}
\examples{
warn_duplicated_treeid <- flag_duplicated_var(warning, treeid)
tree <- tibble::tibble(treeID = c(1, 1))
warn_duplicated_treeid(tree)
flag_duplicated_var(message, treeid)(tree)
}
\seealso{
Other function factories.: \code{\link{detect_duplicated_var}}

Other functions for developers.: \code{\link{detect_duplicated_var}},
  \code{\link{find_data_of_class}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}},
  \code{\link{type_ensure}}
}
\concept{function factories.}
\concept{functions for developers.}
