% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_if_group.R
\name{flag_if_group}
\alias{flag_if_group}
\title{Detect and flag based on a predicate applyed to a variable by groups.}
\usage{
flag_if_group(.data, name, predicate, condition = warn, msg = NULL)
}
\arguments{
\item{.data}{A dataframe.}

\item{name}{String. The name of a column of the dataframe.}

\item{predicate}{A predicate function, e.g. \code{\link[fgeo.base:is_multiple]{fgeo.base::is_multiple()}}.}

\item{condition}{A condition function, e.g. \code{\link[rlang:inform]{rlang::inform()}} or
\code{\link[base:stop]{base::stop()}}.}

\item{msg}{String. An optional custom-messsage.}
}
\value{
\itemize{
\item \code{flag_if_group()}: A condition and its first input, invisibly.
\item \code{detect_if_group()}: Logical of length 1.
}
}
\description{
These functions extend \code{\link[fgeo.base:flag_if]{fgeo.base::flag_if()}} and \code{\link[fgeo.base:detect_if]{fgeo.base::detect_if()}} to
work by groups defined with \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\seealso{
Other functions to check inputs.: \code{\link{detect_duplicated_by_group_f}}

Other functions for developers.: \code{\link{detect_duplicated_by_group_f}},
  \code{\link{find_data_of_class}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}},
  \code{\link{type_ensure}}

Other predicates.: \code{\link{detect_duplicated_by_group_f}}
}
\concept{functions for developers.}
\concept{functions to check inputs.}
\concept{predicates.}
