% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var-ctfs.R
\name{from_var_to_var}
\alias{from_var_to_var}
\alias{rowcol_to_index}
\alias{index_to_rowcol}
\alias{gxgy_to_index}
\alias{gxgy_to_lxly}
\alias{gxgy_to_rowcol}
\alias{gxgy_to_hectindex}
\title{Functions to get variables from other variables.}
\usage{
rowcol_to_index(rowno, colno, gridsize, plotdim)

index_to_rowcol(index, gridsize, plotdim)

gxgy_to_index(gx, gy, gridsize, plotdim)

gxgy_to_lxly(gx, gy, gridsize, plotdim)

gxgy_to_rowcol(gx, gy, gridsize, plotdim)

gxgy_to_hectindex(gx, gy, plotdim)
}
\arguments{
\item{rowno, colno}{Row and column number -- as defined in a census plot.}

\item{gridsize}{The gridsize of the census plot (commonly 20 m).}

\item{plotdim}{The global dimensions of the census plot (i.e. the
maximum possible values of `gx` and `gy`).}

\item{index}{Index number as defined for a census plot.}

\item{gx, gy}{A number; golbal x and y position in a census plot.}
}
\value{
A vector or dataframe (see examples).
}
\description{
These functios wrap their corresponding functions from the CTFS R Package,
but these versions are stricter. The main differences are these:
\itemize{
\item names use "_" not ".".
\item argument gridsize defaults to missing to force the user to provide it.
\item If the argument \code{plotdim} is missing from functions \code{gxgy_*()}, its value
will be guessed and notified.
}
}
\details{
Best is to avoid these functions because some of them output a vector and
others output a dataframe. The recomended alternative to all these functions
is the wrapper \code{\link[=add_var]{add_var()}}, which consistently inputs and outputs a
dataframe, and is formally tested.
}
\examples{
\dontrun{
gxgy_to_index(c(0, 400, 990), c(0, 200, 490), gridsize = 20)

gridsize <- 20
plotdim <- c(1000, 500)

x <- gxgy_to_hectindex(1:3, 1:3, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_index(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_lxly(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_rowcol(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- index_to_rowcol(1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- rowcol_to_index(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)
}
}
\seealso{
\code{\link[=add_var]{add_var()}}.
}
\keyword{internal}
