% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{guess_plotdim}
\alias{guess_plotdim}
\title{Guess plot dimensions.}
\usage{
guess_plotdim(x, accuracy = 50)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{accuracy}{A number giving the accuracy with which to round \code{gx} and
\code{gy}.}
}
\value{
A numeric vector of length 2.
}
\description{
Guess plot dimensions.
}
\examples{
x <- data.frame(
  gx = c(0, 300, 980),
  gy = c(0, 300, 499)
)
guess_plotdim(x)
}
