% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_duplicated.R
\name{is_multiple}
\alias{is_multiple}
\alias{is_duplicated}
\title{Predicates to detect and flag duplicated and multiple values of a variable.}
\usage{
is_multiple(.data)

is_duplicated(.data)
}
\arguments{
\item{.data}{A vector.}
}
\value{
Logical.
}
\description{
\code{is_multiple()} and \code{is_duplicated()} return \code{TRUE} if they detect,
respectively, multiple different values of a variable (e.g. c(1, 2)\code{), or duplicated values of a variable (e.g. c(1, 1)}).
}
\examples{
is_multiple(c(1, 2))
is_multiple(c(1, 1))
is_multiple(c(1, NA))

is_duplicated(c(1, 2))
is_duplicated(c(1, 1))
is_duplicated(c(1, NA))
}
\seealso{
Other functions to check inputs.: \code{\link{flag_if_group}}

Other functions for developers: \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}}

Other functions for internal use in other fgeo packages: \code{\link{flag_if}},
  \code{\link{name_dfs}}, \code{\link{suffix_match}}
}
\concept{functions for developers}
\concept{functions for internal use in other fgeo packages}
\concept{functions to check inputs.}
\concept{general predicates}
\keyword{internal}
