% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{is_named}
\alias{is_named}
\title{Is named object? (Base-only version of rlang::is_named).}
\usage{
is_named(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
The scalar \code{TRUE} or \code{FALSE}.
}
\description{
Is named object? (Base-only version of rlang::is_named).
}
\examples{
is_named(c(x = 1))
is_named(c(1))
is_named(x = 1)

is_named(data.frame(x = 1))

dfm <- data.frame(1)
is_named(dfm)
dfm
}
\seealso{
Other general predicates: \code{\link{is_multiple}}

Other functions for internal use in other fgeo packages: \code{\link{flag_if}},
  \code{\link{guess_plotdim}}, \code{\link{is_multiple}},
  \code{\link{name_dfs}}, \code{\link{suffix_match}}
}
\concept{functions for internal use in other fgeo packages}
\concept{general predicates}
\keyword{internal}
