% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_dfs_to_output.R
\name{list_csv}
\alias{list_csv}
\title{Export mapping each dataframe in a list to a single .csv file.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
list_csv(dfs, dir, prefix = NULL)
}
\arguments{
\item{dfs}{A list of dataframes.}

\item{dir}{Character; the directory where the files will be saved.}

\item{prefix}{Character; a prefix to add to the file names.}
}
\description{
A useful complement of this function is \code{\link[=xlsheets_list]{xlsheets_list()}}.
}
\examples{
dfs <- xlsheets_list(tool_example("multiple_sheets.xlsx"))

# Saving the output to a temporary file
output <- tempdir()
list_csv(dfs, output, prefix = "myfile-")

# Look inside the output directory to confirm it worked
dir(output, pattern = "myfile")
}
\seealso{
xlsheets_list

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{xlsheets_list}}

Other general functions to export data: \code{\link{list_df}}
}
\concept{functions to handle multiple spreadsheets of an excel workbook.}
\concept{general functions to export data}
