% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_manipulate_lists.R
\name{ls_csv_df}
\alias{ls_csv_df}
\title{Save each dataframe in a list to a different .csv file.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
ls_csv_df(df_list, dir, prefix = NULL)
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{dir}{Character; the directory where the files will be saved.}

\item{prefix}{Character; a prefix to add to the file names.}
}
\description{
A useful complement of this function is \code{\link[=ls_list_spreadsheets]{ls_list_spreadsheets()}}.
}
\examples{
path_to_excel_workbook <- system.file(
  "extdata", "example.xlsx", package = "fgeo.tool"
)

df_list <- ls_list_spreadsheets(path_to_excel_workbook)
str(df_list)

output <- tempdir()
ls_csv_df(df_list, output, prefix = "myfile-")

files <- dir(output)
files[grepl(".csv$", files)]
}
\seealso{
ls_list_spreadsheets.

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{ls_list_spreadsheets}}
}
\concept{functions to handle multiple spreadsheets of an excel workbook.}
