% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_manipulate_lists.R
\name{ls_list_spreadsheets}
\alias{ls_list_spreadsheets}
\title{Read an excel workbook and store each spreadsheet as a dataframe in a list.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
ls_list_spreadsheets(path)
}
\arguments{
\item{path}{A path to an excel file.}
}
\value{
A list of dataframes.
}
\description{
A useful complement of this function is \code{\link[=ls_csv_df]{ls_csv_df()}}.
}
\examples{
path_to_excel_workbook <- system.file(
  "extdata", "example.xlsx", package = "fgeo.tool"
)
str(ls_list_spreadsheets(path_to_excel_workbook))
}
\seealso{
ls_csv_df

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{ls_csv_df}}
}
