% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_manipulate_lists.R
\name{ls_name_df}
\alias{ls_name_df}
\title{Mutate a list of dataframes to add column with the name of the list element.}
\usage{
ls_name_df(df_list, name = "name")
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{name}{A string giving the name of the column that stores the name of
each dataframe of \code{x}.}
}
\value{
A modified version of \code{x}.
}
\description{
Mutate a list of dataframes to add column with the name of the list element.
}
\examples{
df_list <- list(a = data.frame(a = 1), data.frame(a = 2, b = 2))
df_list
ls_name_df(df_list)
ls_name_df(df_list, name = "sheet")
}
