% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_var.R
\name{multiple_var_by_group}
\alias{multiple_var_by_group}
\title{Factory of functions to detect multiple values of a variable by groups.}
\usage{
multiple_var_by_group(var)
}
\arguments{
\item{.data}{}
}
\description{
Extends fgeo.base::multiple_var().
}
