% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_dfs.R
\name{name_dfs}
\alias{name_dfs}
\title{Identify each dataframe in a list with a name.}
\usage{
name_dfs(dfs, name = "name")
}
\arguments{
\item{dfs}{A list of dataframes.}

\item{name}{Names of the columns that store the names and values.}
}
\value{
A list of dataframes.
}
\description{
Identify each dataframe in a list with the name of the corresponding list
item.
}
\examples{
dfs <- list(a = data.frame(x = 1), b = data.frame(x = 1))
name_dfs(dfs)

name_dfs(dfs, "custom_name")

dfs2 <- list(data.frame(x = 1), data.frame(x = 1))
name_dfs(dfs2)
}
\seealso{
Other functions for internal use in other fgeo packages: \code{\link{flag_if}},
  \code{\link{is_multiple}}, \code{\link{suffix_match}}
}
\concept{functions for internal use in other fgeo packages}
\concept{general functions to deal with names}
\keyword{internal}
