% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_lowercase.R
\name{names_anycase}
\alias{names_anycase}
\title{Find a name exactly matching a string but regardless of case.}
\usage{
names_anycase(x, nm)
}
\arguments{
\item{x}{A named object.}

\item{nm}{A string to match names exactly but regardless of case.}
}
\value{
A string of the name that was found in \code{names(x)}.
}
\description{
Find a name exactly matching a string but regardless of case.
}
\examples{
v <- c(a = 1, B = 1)
names_anycase(v, "b")

dfm <- data.frame(a = 1, B = 1)
names_anycase(dfm, "b")
}
