% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_lowercase.R
\name{names_lowercase}
\alias{names_lowercase}
\alias{names_lowercase}
\alias{names_restore}
\title{Lower and restore names.}
\usage{
names_lowercase(x)

names_restore(x)
}
\arguments{
\item{x}{A named object.}
}
\value{
\itemize{
\item \code{names_lowercase()} Returns the object \code{x} with lowercase names
\item \code{names_restore()} Returns the object \code{x} with original (restored) names.
}
}
\description{
These functions are useful together, to lowercase names, then to do something
(as long as attributes are preserved -- see section warning), and then
restore the original names.
}
\section{Warning}{

\code{\link[=names_restore]{names_restore()}} is similar to \code{\link[=names_restore_new_var]{names_restore_new_var()}} but
\code{\link[=names_restore]{names_restore()}} is necesary if the data is mutated with \code{\link[dplyr:mutate]{dplyr::mutate()}}:
\code{\link[dplyr:mutate]{dplyr::mutate()}} drops attributes
(https://github.com/tidyverse/dplyr/issues/1984), which makes it
\code{\link[=names_restore]{names_restore()}} useless. attributes.
}

\examples{
cns <- tibble::tibble(CensusID = 1, status = "A")
original <- cns
original

lowered <- names_lowercase(cns)
lowered
attr(lowered, "names_old")

back_to_original <- names_restore(lowered)
back_to_original
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{exists_in_pkg}},
  \code{\link{fgeo_package_deps}},
  \code{\link{names_restore_new_var}}
}
