% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_manipulate_names.R
\name{nms}
\alias{nms}
\alias{nms_has_any}
\alias{nms_detect}
\alias{nms_extract_all}
\alias{nms_extract1}
\title{Functions to detect and extract names.}
\usage{
nms_has_any(x, ...)

nms_detect(x, ...)

nms_extract_all(x, ...)

nms_extract1(x, ...)
}
\arguments{
\item{x}{A named object.}

\item{...}{Strings; each of the names that need to be checked.}
}
\description{
These functions are handy to work with fgeo's data structures because the
same variable may be named differently in different data sets. For example,
the variable status is called \code{Status} or \code{status} in viewfull or census
(tree and stem) tables.
}
\details{
nms_has_any(): Checks if an object has any of the provided names.
\itemize{
\item Returns a logical value.
nms_detect(): Checks if an object has the provided names.
\item Returns a logical vector.
nms_extract_all(): Finds the names that match the provided names.
\item Returns a character vector.
nms_extract1(): Finds the first name that matches the provided names.
\item Returns a character string.
}
}
\examples{
v <- c(a = 1, b = 1)
nms_has_any(v, "a", "B")
nms_has_any(v, "A", "B")
nms_has_any(v, "A", "b")

nms_detect(v, "a", "B", "b")

nms_extract_all(v, "a", "B")
nms_extract_all(v, "a", "a", "b")

nms_extract1(v, "a", "a", "b")

# Usage with ForestGEO data

vft <- fgeo.x::vft_4quad
nms_extract_all(vft, "gx", "gy", "PX", "PY")

stem <- fgeo.x::stem6
nms_extract_all(stem, "gx", "gy", "PX", "PY")
}
\seealso{
Other functions for developers: \code{\link{check_crucial_names}},
  \code{\link{extract_insensitive}},
  \code{\link{find_data_of_class}},
  \code{\link{flag_if_group}}, \code{\link{is_multiple}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}},
  \code{\link{rename_matches}}, \code{\link{type_ensure}}

Other functions dealing with names: \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}
}
\concept{functions dealing with names}
\concept{functions for developers}
\keyword{internal}
