% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_manipulate_names.R
\name{nms_extract_anycase}
\alias{nms_extract_anycase}
\title{Find a name exactly matching a string but regardless of case.}
\usage{
nms_extract_anycase(x, nm)
}
\arguments{
\item{x}{A named object.}

\item{nm}{A string to match names exactly but regardless of case.}
}
\value{
A string of the name that was found in \code{names(x)}.
}
\description{
Find a name exactly matching a string but regardless of case.
}
\examples{
v <- c(a = 1, B = 1)
nms_extract_anycase(v, "b")

dfm <- data.frame(a = 1, B = 1)
nms_extract_anycase(dfm, "b")
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{ensure_type}},
  \code{\link{exists_in_pkg}}, \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}}

Other functions to manipulate names.: \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}}
}
