% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_manipulate_names.R
\name{nms_lowercase}
\alias{nms_lowercase}
\alias{nms_lowercase}
\alias{nms_restore}
\title{Lower and restore names.}
\usage{
nms_lowercase(x)

nms_restore(x)
}
\arguments{
\item{x}{A named object.}
}
\value{
\itemize{
\item \code{nms_lowercase()} Returns the object \code{x} with lowercase names
\item \code{nms_restore()} Returns the object \code{x} with original (restored) names.
}
}
\description{
These functions are useful together, to lowercase names, then to do something
(as long as attributes are preserved -- see section warning), and then
restore the original names.
}
\section{Warning}{

\code{\link[=nms_restore]{nms_restore()}} is similar to \code{\link[=nms_restore_newvar]{nms_restore_newvar()}} but
\code{\link[=nms_restore]{nms_restore()}} is necessary if the data is mutated with \code{\link[dplyr:mutate]{dplyr::mutate()}}:
\code{\link[dplyr:mutate]{dplyr::mutate()}} drops attributes
(https://github.com/tidyverse/dplyr/issues/1984), which makes it
\code{\link[=nms_restore]{nms_restore()}} useless. attributes.
}

\examples{
cns <- tibble::tibble(CensusID = 1, status = "A")
original <- cns
original

lowered <- nms_lowercase(cns)
lowered
attr(lowered, "names_old")

back_to_original <- nms_restore(lowered)
back_to_original
}
\seealso{
Other functions for developers.: \code{\link{exists_in_pkg}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}},
  \code{\link{type_ensure}}

Other functions to manipulate names.: \code{\link{nms_extract_anycase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}}
}
