% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{nms_tidy}
\alias{nms_tidy}
\alias{nms_tidy}
\alias{as_tidy_names}
\title{Create tidy names that are lowercase and have no empty spaces.}
\usage{
nms_tidy(x)

as_tidy_names(x)
}
\arguments{
\item{x}{A named object or a character string.}
}
\value{
A modified version of \code{x} with tidy names or a string of tidy names.
}
\description{
These functions create tidy names that are lowercase and have no empty
spaces. \code{nms_tidy()} tidyes the names of named objects. Unnamed strings are
also tidied via \code{as_tidy_names()}. \code{as_tidy_names()} tidies a string -- not
its names.
}
\examples{
messy <- "Hi mE"

nms_tidy(messy)
# Same
as_tidy_names(messy)

messy_df <- setNames(data.frame(1), messy)
messy_df
nms_tidy(messy_df)

messy_list <- setNames(list(1), messy)
messy_list
nms_tidy(messy_list)

messy_vector <- setNames(c(1), messy)
messy_vector
nms_tidy(messy_vector)
}
