% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_dbh_largest.R
\name{pick_dbh_largest}
\alias{pick_dbh_largest}
\title{Pick the (stem of) largest dbh per treeid per censusid.}
\usage{
pick_dbh_largest(.x)
}
\arguments{
\item{.x}{A ForestGEO-like dataframe, census or ViewFullTable.}
}
\value{
A dataframe with one row per per treeid per censusid and a single
plotname.
}
\description{
This function picks the stem with the maximum dbh per treeid per censusid. It
intentionally errs if it detects multiple distinct values of plotname (i.e.
if the data is a ViewFullTable with multiple plots).
}
\examples{
census <- tibble::tribble(
  ~dbh,   ~sp, ~treeID, ~stemID,
    10, "sp1",     "1",   "1.1",
   100, "sp1",     "1",   "1.2",
    22, "sp2",     "2",   "2.1",
    99, "sp2",     "2",   "2.2",
    99, "sp2",     "2",   "2.3",
    NA, "sp2",     "2",   "2.4"
)

pick_dbh_largest(census)
}
\seealso{
Other functions to pick or drop rows of a dataframe.: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}}, \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{functions to pick or drop rows of a dataframe.}
