% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_largest_hom_dbh.R
\name{pick_largest_hom_dbh}
\alias{pick_largest_hom_dbh}
\title{Pick the main stem of each tree in each census.}
\usage{
pick_largest_hom_dbh(.x)
}
\arguments{
\item{.x}{A ForestGEO-like dataframe, census or ViewFullTable.}
}
\value{
A dataframe with one row per per treeid per censusid and a single
plotname.
}
\description{
This function picks the main stem of each tree in each census. It
intentionally errs if it detects multiple plots. It ignores groups of grouped
data -- to operate within groups see ?\code{\link[=pick_woods]{pick_woods()}}.
}
\examples{
census <- tibble::tribble(
  ~dbh,   ~sp, ~treeID, ~stemID, ~hom,
    10, "sp1",     "1",   "1.1",   10,
   100, "sp1",     "1",   "1.2",   10,
    22, "sp2",     "2",   "2.1",   10,
    99, "sp2",     "2",   "2.2",   10,
    99, "sp2",     "2",   "2.3",   10,
    NA, "sp2",     "2",   "2.4",   10
)

pick_largest_hom_dbh(census)
}
\seealso{
Other functions to pick or drop rows of a dataframe.: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}}, \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{functions to pick or drop rows of a dataframe.}
