% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_woods.R
\name{pick_woods_f}
\alias{pick_woods_f}
\title{Factory of functions to pick woods and then apply a function.}
\usage{
pick_woods_f(.f, .collapse = fgeo.tool::pick_largest_hom_dbh)
}
\arguments{
\item{.f}{A function to apply after picking woods. For example, \code{.f = identity} returns the picked woods unchanged.}

\item{.collapse}{A function to collapse multiple values of \code{treeID} into a
single one.}
}
\description{
Factory of functions to pick woods and then apply a function.
}
