% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tbl.R
\name{print.tbl}
\alias{print.tbl}
\alias{print.tbl_df}
\title{Print all rows and/or columns of a tibble.}
\usage{
\method{print}{tbl}(x, ..., n = NULL, width = NULL, n_extra = NULL)

\method{print}{tbl_df}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)
}
\arguments{
\item{x}{Object to show.}

\item{...}{Other arguments passed to \code{format()}.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print
all rows if less than option \code{tibble.print_max}. Otherwise, will
print \code{tibble.print_min} rows.}

\item{width}{Width of text output to generate. This defaults to NULL, which
means use \code{getOption("tibble.width")} or (if also NULL)
\code{getOption("width")}; the latter displays only the columns that
fit on one screen. You can also set \code{options(tibble.width = Inf)} to
override this default and always print all columns.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the
default, will print information about at most \code{tibble.max_extra_cols}
extra columns.}
}
\value{
Print output (and the main input invisibly).
}
\description{
Print all rows and/or columns of a tibble.
}
\examples{
print(as_tibble(iris), n = 1)
print(as_tibble(iris), n = 100)
print(fgeo.data::luquillo_tree5_random)
print(fgeo.data::luquillo_tree5_random, n_extra = 2)
print(fgeo.data::luquillo_tree5_random, width = Inf)
}
\concept{general miscellaneous functions}
