% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_df.R
\name{rdata_df}
\alias{rdata_df}
\title{Import multiple .rdata files from a directory into a dataframe.}
\usage{
rdata_df(path_dir, .match = NULL, .id = NULL)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{.match}{If not \code{NULL} a string to match specific .rdata file-names
(excluding the extension).}

\item{.id}{If not NULL a variable with this name will be created to inform
the source file.}
}
\value{
A dataframe.
}
\description{
This is shortcut for \code{list_df(rdata_list(x))} with the additional ability to
read exclusively .rdata files containing a specific string. It is designed to
expect that all .rdata files contain dataframes with the same names -- as is
the case for multiple ForestGEO censuses delivered by the database. Notice
that function name have the format input_output, i.e. rdata_df.
}
\examples{
path_dir <- tool_example("rdata")
dir(path_dir)

rdata_df(path_dir)

rdata_df(path_dir, .match = "6")

dfm <- rdata_df(path_dir, .match = "tree5|6", .id = "source")
dfm
tail(dfm)
}
\seealso{
Other general functions to import data: \code{\link{dir_list}},
  \code{\link{read_with}}, \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
