% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_censuses_df.R
\name{read_censuses_df}
\alias{read_censuses_df}
\title{Read ForestGEO censuses.}
\usage{
read_censuses_df(path_dir, .match = NULL, .id = "censuses")
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{.match}{If not \code{NULL} a string to match specific .rdata file-names
(excluding the extension).}

\item{.id}{If not NULL a variable with this name will be created to inform
the source file.}
}
\value{
An object of class 'censuses_df'.
}
\description{
This function reads any number of ForestGEO censuses stored in .rdata files
and stores them in a nested dataframe of class 'censuses_df', which has a
convenient print method and enables using specific methods for working with
ForestGEO data.
}
\examples{
censuses <- read_censuses_df(tool_example("rdata"))
class(censuses)
censuses$data

pick(censuses, dbh > 50)$data
}
\seealso{
\code{\link[=rdata_df]{rdata_df()}}, \code{\link[=as_censuses]{as_censuses()}}.

Other functions to import ForestGEO data: \code{\link{read_fgeo}}
}
\concept{functions to import ForestGEO data}
