% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_list.R
\name{read_with}
\alias{read_with}
\title{Import multiple files into a list using any given reader function.}
\usage{
read_with(.f, regexp = NULL)
}
\arguments{
\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression to match files of appropriate format.}
}
\value{
A modified version of the input function, able to read all files
from a directory (provided they all are of the suitable format).
}
\description{
Import multiple files into a list using any given reader function.
}
\examples{
path_rds <- tool_example("rds")
path_rds
dir(path_rds)

rds_list <- read_with(readr::read_rds)
rds_list(path_rds)

# Same
read_with(readr::read_rds)(path_rds)
}
\seealso{
Other general functions to import data: \code{\link{dir_list}},
  \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
