% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subquad.R
\name{recode_subquad}
\alias{recode_subquad}
\title{Recode subquadrat.}
\usage{
recode_subquad(x, offset = -1)
}
\arguments{
\item{x}{A dataframe with the variable \code{subquadrat}.}

\item{offset}{A number; either -1 or 1, to rest or add one unit to the number
of column of each subquadrat.\preformatted{First column is 0    First column is 1
-----------------    -----------------
   04 14 24 34          14 24 34 44
   03 13 23 33          13 23 33 43
   02 12 22 32          12 22 32 42
   01 11 21 31          11 21 31 41
}}
}
\value{
A modified version of the input.
}
\description{
Recode subquadrat.
}
\examples{
first_subquad_11 <- tibble::tibble(subquadrat = c("11", "12", "22"))
first_subquad_11

first_subquad_01 <- recode_subquad(first_subquad_11, offset = -1)
first_subquad_01

first_subquad_11 <- recode_subquad(first_subquad_01, offset = 1)
first_subquad_11
}
\keyword{internal}
