% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_all_na.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace all NA's of a dataframe with a filler}
\usage{
replace_all_na(df, filler = 0)
}
\arguments{
\item{df}{A dataframe.}

\item{replace}{A filler; whatever you want to replace NA's with.}
}
\value{
A modified versio of \code{df}.
}
\description{
Replace all NA's of a dataframe with a filler
}
\examples{
df <- data.frame(x = c(NA, 1), y = c("a", NA), stringsAsFactors = FALSE)
replace_all_na(df)
replace_all_na(df, "")
replace_all_na(df, "missing")
}
