% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_censusid.R
\name{row_collapse_censusid}
\alias{row_collapse_censusid}
\title{Collapse the values of CensusID.}
\usage{
row_collapse_censusid(x)
}
\arguments{
\item{x}{A dataframe -- specifically a ViewFullTable.}
}
\value{
A modified version of \code{x}, most likely with less rows.
}
\description{
This function is useful to summarize the status history of a tree tag accross
censuses. It is particularly useful to avoid duplicated tree tags on maps.
}
\examples{
x <- tibble::tribble(
  ~CensusID, ~Tag, ~Status,
  "1",  "1", "alive",
  "1",  "1",  "dead",
  "1",  "2",  "dead",
  "1",  "2",  "dead",
  "2",  "1", "alive",
  "2",  "1", "alive",
  "2",  "2", "alive",
  "2",  "2",  "dead"
)

row_collapse_censusid(x)

with_status_tree <- fgeo.tool::add_status_tree(x, "alive", "dead")
reduced <- unique(dplyr::select(with_status_tree, -Status))
reduced

row_collapse_censusid(reduced)
}
