% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_top.R
\name{row_top}
\alias{row_top}
\title{Filter a data set by matching n values from the head (or tail) of a variable.}
\usage{
row_top(.data, var, n = 1)
}
\arguments{
\item{.data}{A table of data}

\item{var}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right.
}

The default returns the last column (on the assumption that's the
column you've created most recently).

This argument is taken by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names and column positions).}

\item{n}{Number of values used for matching, from the head (or tail) of \code{var}.}
}
\value{
A filtered version of the input dataset.
}
\description{
Filters a data frame by matching \code{n} values from the head (or tail) of the
unique values of a variable.
}
\details{
Similar to \code{\link[dplyr:top_n]{dplyr::top_n()}}, but instead of using \code{min_rank()} or
\code{max_rank()}, it uses \code{\link[utils:head]{utils::head()}} or \code{\link[utils:tail]{utils::tail()}}; and \code{var} is
flexible as in \code{\link[dplyr:pull]{dplyr::pull()}}.
}
\examples{
df <- data.frame(x = 1:9, y = letters[1:3], stringsAsFactors = FALSE)

# `var` can be bare or quoted
(result <- row_top(df, "y"))
identical(row_top(df, y), result)

# matching `var` by position starting from the left
identical(row_top(df, var = y), row_top(df, var = 2))
# matching `var` by position starting from the right
identical(row_top(df, var = y), row_top(df, var = -1))

row_top(df, y, n = 2)
# Negative values select from the tail
row_top(df, y, n = -2)
}
\seealso{
\link[dplyr:pull]{dplyr::pull}, \link[dplyr:top_n]{dplyr::top_n}, \code{\link[utils:head]{utils::head()}}, \code{\link[utils:tail]{utils::tail()}}.

Other functions to manipulate dataframe rows.: \code{\link{row_discard_twice_dead}},
  \code{\link{row_filter_status}}
}
