% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize}
\alias{sanitize}
\alias{sanitize_vft}
\alias{sanitize_taxa}
\title{Sanitize ViewFullTable and ViewTaxonomy.}
\usage{
sanitize_vft(x, na = c("", "NA", "NULL"), ...)

sanitize_taxa(x, na = c("", "NA", "NULL"), ...)
}
\arguments{
\item{x}{A dataframe; either a ForestGEO ViewFullTable (\code{sanitize_vft()}) or
ViewTaxonomy (\code{sanitize_vft()}).}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{...}{Arguments passed on to \code{readr::type_convert}
\describe{
  \item{df}{A data frame.}
  \item{col_types}{One of \code{NULL}, a \code{\link[=cols]{cols()}} specification, or
a string. See \code{vignette("column-types")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[=cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[=cols_only]{cols_only()}}.

Unlike other functions \code{type_convert()} does not allow character
specifications of \code{col_types}.}
  \item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}
  \item{trim_ws}{Should leading and trailing whitespace be trimmed from
each field before parsing it?}
  \item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}}
}
\value{
A dataframe.
}
\description{
These functions sanitize dataframes from ViewFullTable and ViewTaxonomy to
achieve the same effect as reading with \code{read_vft()} and \code{read_taxa()}. The
most common problems that these functions fix are these:
\itemize{
\item Ensure that each column has the correct type.
\item Ensure that missing values are represented with \code{NA}s -- not with the
literal string "NULL".
}
}
\section{Acknowledgments}{

Thanks to Shameema Jafferjee Esufali for motivating this functions.
}

\examples{
vft <- fgeo.data::luquillo_vft_4quad

# Introdue problems to show how to fix them
# Bad column types
vft[] <- lapply(vft, as.character)
# Bad representation of missing values
vft$PlotName <- "NULL"

# "NULL" should be replaced by `NA` and `DBH` should be numeric
str(vft[c("PlotName", "DBH")])

# Fix
vft_sane <- sanitize_vft(vft)
str(vft_sane[c("PlotName", "DBH")])



taxa <- fgeo.data::luquillo_taxa
# Bad column types
taxa[] <- lapply(taxa, as.character)
# Bad representation of missing values
taxa$SubspeciesID <- "NULL"

# "NULL" should be replaced by `NA` and `ViewID` should be integer
str(taxa[c("SubspeciesID", "ViewID")])

# Fix
taxa_sane <- sanitize_taxa(taxa)
str(taxa_sane[c("SubspeciesID", "ViewID")])
}
\seealso{
\code{\link[=read_fgeo]{read_fgeo()}}.
}
\concept{functions to edit ForestGEO data in place}
