% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitze_vft}
\alias{sanitze_vft}
\title{Sanitize ViewFullTable and ViewTaxonomy.}
\usage{
sanitze_vft(vft)
}
\arguments{
\item{vft}{A dataframe; particularly a ForestGEO ViewFullTable (\code{vft}).}

\item{taxa}{A dataframe; particularly a ForestGEO ViewTaxonomy (\code{taxa}).}
}
\description{
These functions sanitize dataframes from ViewFullTable and ViewTaxonomy to
achieve the same effect as reading with \code{read_vft()} and \code{read_taxa()}. The
most common problems that these functions fix are these:
\itemize{
\item Ensure that each column has the correct type.
\item Ensure that missing values are represented with \code{NA}s -- not with the
literal string "NULL".
}
}
\section{Acknowledgments}{

Thanks to Shameema Jafferjee Esufali for motivating this functions.
}

\examples{
# Example data
vft <- tibble::tibble(QuadratName = c(0000, 0001), DBH = c("NULL", "NULL"))
str(vft)

# Warns because lot's of columns are missing
sane <- sanitze_vft(vft)
str(sane, give.attr = FALSE)

taxa <- luquillo::ViewTaxonomy_luquillo
# Introduce mistakes
taxa$SubspeciesID <- "NULL"
taxa$SpeciesID <- as.double(taxa$SpeciesID)
str(taxa[1:3], give.attr = FALSE)

sane <- sanitze_taxa(taxa)
str(sane[1:3], give.attr = FALSE)
}
\seealso{
\code{\link[=read_fgeo]{read_fgeo()}}.
}
