% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{standardize_at}
\alias{standardize_at}
\title{Standardize specific columns of a dataframe.}
\usage{
standardize_at(x, .at, total)
}
\arguments{
\item{x}{Dataframe.}

\item{total}{A numeric vector of length 1.}
}
\value{
A data.frame.
}
\description{
Standardize specific columns of a dataframe.
}
\examples{
dfm <- data.frame(a = 1:3, b = 11:13)
standardize_at(dfm, "a", total = 100)
standardize_at(dfm, names(dfm), total = 100)
}
