% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_status.R, R/stem_status.R
\name{stem_not_dead}
\alias{stem_not_dead}
\alias{tree_not_dead}
\alias{status_stem}
\alias{stem_not_dead}
\title{Filter a data set by the status of its stems.}
\usage{
stem_not_dead(x, .status = "D")

tree_not_dead(x, .status = "dead")

status_stem(x, .status, exclude = FALSE)

stem_not_dead(x, .status = "D")
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{.status}{Character vector; Must be one of possible values of the
variable giving the status of the dataframe \code{x}.}

\item{exclude}{Logical; \code{TRUE} excluded the values passed to \code{.status}.}
}
\value{
A filtered version of the dataframe \code{x}.
}
\description{
Filter a data set by the status of its stems.
}
\examples{
x <- tibble::tibble(status = LETTERS[1:4])
x

status_stem(x, "A")
status_stem(x, c("A", "C"))
status_stem(x, c("A", "C"), exclude = TRUE)

status_stem(x, c("D"), exclude = TRUE)
# Shortcut
stem_not_dead(x)
}
