% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_suffix_match.R
\name{str_suffix_match}
\alias{str_suffix_match}
\title{Suffix a strings where a vector exactly matches one specific string.}
\usage{
str_suffix_match(string, to_match, .match, suffix)
}
\arguments{
\item{string}{A vector of strings to suffix.}

\item{to_match}{A vector of strings to match with \code{.match}.}

\item{.match}{A string to match the vector \code{to_match} with.}

\item{suffix}{A suffix to add at the end of each element of \code{string}.}
}
\value{
A modified version of \code{x}.
}
\description{
Suffix a strings where a vector exactly matches one specific string.
}
\examples{
str_suffix_match(
  string = c("tag1", "tag2"),
  to_match = c("dead", "not-dead"),
  .match = "dead",
  suffix = ".d"
)

library(dplyr)
vft <- tribble(
  ~Tag, ~Status,
  "01", "dead",
  "02", "alive"
)
mutate(vft, tagged = str_suffix_match(Tag, Status, "dead", ".d"))
}
