% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.demography_impl}
\alias{to_df.demography_impl}
\title{Dataframe objects of class "demography_impl".}
\usage{
\method{to_df}{demography_impl}(.x, ...)
}
\arguments{
\item{.x}{An object of class demography_impl.}

\item{...}{Other arguments passed to \code{to_df()}.}
}
\value{
A (tibble) dataframe.
}
\description{
Dataframe objects of class "demography_impl".
}
\examples{
library(fgeo.ctfs)

census1 <- fgeo.x::tree5
census2 <- fgeo.x::tree6

by <- interaction(census1$sp, census1$quadrat, sep = "__")
.x <- recruitment_impl(census1, census2, split1 = by)
head(to_df(.x))
}
\seealso{
\code{\link[=to_df]{to_df()}}.

Other methods for fgeo generics: \code{\link{to_df.krig_lst}},
  \code{\link{to_df.tt_lst}}
}
\concept{methods for fgeo generics}
