% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.demography_lst}
\alias{to_df.demography_lst}
\alias{to_df.demography_lst_by}
\title{Dataframe objects of class "demography_lst" and "demography_lst_by".}
\usage{
\method{to_df}{demography_lst}(.x, ...)

\method{to_df}{demography_lst_by}(.x, ...)
}
\arguments{
\item{.x}{An object of class demography_lst.}

\item{...}{Other arguments passed to \code{to_df()}.}
}
\value{
A (tibble) dataframe.
}
\description{
This method creates a dataframe from the output of
\code{fgeo.demography::mortality()}, \code{fgeo.demography::recruitment()}, and
\code{fgeo.demography::growth()} (each one is a list of class "demography_lst" if
\code{by} is \code{NULL}, or of class "demography_lst" if \code{by} is not \code{NULL}):
}
\details{
\itemize{
\item \code{to_df.demography_lst()}: Restructures results calculated across the entire
census data.
\item \code{to_df.demography_lst_by()}: Restructures results calculated \code{by} groups.
}
}
\examples{
\dontrun{
library(fgeo.data)
library(fgeo.demography)

censuses <- list(luquillo_tree5_random, luquillo_tree6_random)
to_df(mortality(censuses))

to_df(mortality(censuses, "sp"))
}
}
\seealso{
\code{\link[=to_df]{to_df()}}.

Other methods for fgeo generics: \code{\link{to_df.krig_lst}},
  \code{\link{to_df.tt_lst}}
}
\concept{methods for fgeo generics}
