% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.krig_lst}
\alias{to_df.krig_lst}
\title{Restructure the output of \code{fgeo.habitat::krig()} as a dataframe.}
\usage{
\method{to_df}{krig_lst}(.x, name = "var", item = "df", ...)
}
\arguments{
\item{.x}{The output of \code{\link[fgeo.habitat:krig]{fgeo.habitat::krig()}}.}

\item{name}{Name for the column to hold soil variable-names.}

\item{item}{Character string; either "df" or "df.poly".}

\item{...}{Arguments passed on to \code{to_df}
\describe{
  \item{.x}{An fgeo object of supported class.}
}}
}
\value{
A dataframe.
}
\description{
Restructure the output of \code{fgeo.habitat::krig()} as a dataframe.
}
\examples{
\dontrun{
library(fgeo.habitat)
vars <- c("c", "p")
krig <- krig(soil_fake, vars, quiet = TRUE)
head(to_df(krig))
}
}
