% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df}
\alias{to_df}
\title{Restructure data as a dataframe.}
\usage{
to_df(.x, ...)
}
\arguments{
\item{.x}{An fgeo object of supported class.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
The goal of this (generic) function (with methods for multiple fgeo classes)
is to produce a dataframe that helps you to work fluently with other general
purpose tools such as \strong{dplyr} and \strong{ggplot2}.
}
\seealso{
\code{\link[=to_df.krig_lst]{to_df.krig_lst()}}, \code{\link[=to_df.tt_lst]{to_df.tt_lst()}}.
}
