% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.tt_lst}
\alias{to_df.tt_lst}
\title{Restructure  the output of \code{tt_test()} as a dataframe.}
\usage{
\method{to_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_lst.}

\item{...}{Other arguments passed to \code{\link[=to_df]{to_df()}}.}
}
\value{
A dataframe.
}
\description{
Restructure  the output of \code{tt_test()} as a dataframe.
}
\examples{
\dontrun{
library(fgeo.habitat)
cns <- luquillo_top3_sp
spp <- unique(cns$sp)
hab <- luquillo_habitat

tt_lst <- tt_test(cns, spp, hab)
tt_df <- to_df(tt_lst)
head(tt_df)
}
}
