% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_example.R
\name{tool_example}
\alias{tool_example}
\title{Path to directory containing example data.}
\usage{
tool_example(path)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/.}
}
\value{
Path to directory containing example data.
}
\description{
Path to directory containing example data.
}
\examples{
tool_example("csv")
dir(tool_example("csv"))
csv_list(tool_example("csv"))
}
\keyword{internal}
