% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_fgeo.R
\name{type_fgeo}
\alias{type_fgeo}
\alias{type_vft}
\alias{type_taxa}
\title{Help read ForestGEO data with consistent column type.}
\usage{
type_vft()

type_taxa()
}
\value{
A list.
}
\description{
This functions helps to read ForestGEO's data with consistent column types.
They output the specification for the argument \code{col_types()} of the functions
\code{read_*()} of the \strong{readr} package:
\itemize{
\item \code{type_vft():} Type specification for ViewFullTable.
\item \code{type_taxa():} Type specification for ViewFullTaxonomy.
}
}
\examples{
type_vft()
type_taxa()
}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}}.
}
