% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_treeid.R
\name{warn_duplicated_treeid}
\alias{warn_duplicated_treeid}
\alias{detect_duplicated_treeid}
\title{Detect and warn duplicated treeid.}
\usage{
warn_duplicated_treeid(.data)

detect_duplicated_treeid(.data)
}
\arguments{
\item{.data}{A ForestGEO dataframe -- census or ViewFullTable.}
}
\value{
\itemize{
\item \code{warn_duplicated_treeid()}: A warning or invisible \code{.data}.
\item \code{detect_duplicated_treeid()}: \code{TRUE} or \code{FALSE}.
}
}
\description{
Detect and warn duplicated treeid.
}
\examples{
stem <- tibble(treeID = c(1, 1), stemID = c(1.1, 1.2))
warn_duplicated_treeid(stem)

# No duplicates
stem <- tibble(treeID = c(1, 2), stemID = c(1.1, 2.1))
warn_duplicated_treeid(tree)

vft <- tibble(TreeID = c(1, 2))
detect_duplicated_treeid(vft)
vft <- tibble(TreeID = c(1, 1))
detect_duplicated_treeid(vft)

# Warns even if the duplicatd treeid come from multiple censuses.
tree <- tibble(CensusID = c(1, 2), treeID = c(1, 1))
detect_duplicated_treeid(tree)
# Deal with this separately, e.g.
split_censusid <- split(tree, tree$CensusID)
any(purrr::map_lgl(split_censusid, detect_duplicated_treeid))
}
