% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listed_df.R
\name{xl_list_sheets}
\alias{xl_list_sheets}
\title{Read an excel workbook and store each spreadsheet as a dataframe in a list.}
\source{
Adappted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
xl_list_sheets(path)
}
\arguments{
\item{path}{A path to an excel file.}
}
\value{
A list of dataframes.
}
\description{
A useful complement of this function is \code{\link[=csv_listed_df]{csv_listed_df()}}.
}
\examples{
path_to_excel_workbook <- system.file(
  "extdata", "example.xlsx", package = "qcr"
)
str(xl_list_sheets(path_to_excel_workbook))
}
\seealso{
csv_listed_df

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{csv_listed_df}}
}
