% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_sheets_to_output.R
\name{xl_sheets_to_output}
\alias{xl_sheets_to_output}
\alias{xl_sheets_to_csv}
\alias{xl_sheets_to_xl}
\alias{xl_sheets_to_df}
\title{Combine spreadsheets from excel workbooks and output common data structures.}
\usage{
xl_sheets_to_csv(input_dir, output_dir = "./")

xl_sheets_to_xl(input_dir, output_dir = "./")

xl_sheets_to_df(input_dir)
}
\arguments{
\item{input_dir}{String giving the directory containing the excel workbooks
to read from.}

\item{output_dir}{String giving the directory where to write .csv files to.}
}
\value{
Writes one .csv file for each workbook.
}
\description{
These functions combine spreadsheets from excel workbooks into common data
structures. \code{xl_sheets_to_csv()} and \code{xl_sheets_to_xl()} write a .csv or
excel (.xlsx) file per workbook -- combining all spreadsheets.
\code{xl_sheets_to_df} outputs a list where each dataframes combines all
spreadsheeets of a workbook.
}
\details{
This is a rigid function with a very specific goal: To process data from a
specific sampling software -- FastField. Specifically, this is what this
function does:
\itemize{
\item Reads each spreadsheet from each workbook and map it to a dataframe.
\item Lowercases and links the names of each dataframe.
\item Keeps only these dataframes: (1) original_stems; (2) new_secondary_stems;
and (3).
\item Dates the data by \code{submission_id} (\code{date} comes from the spreadsheet
\code{root}).
\item Lowercases and links the names of each dataframe-variable.
\item Drops fake stems.
\item Output a common data structure of your choice.
}
}
\examples{
library(fs)

# Path to the folder I want to read excel files from
input_dir <- dirname(example_path("two_files/new_stem_1.xlsx"))
input_dir

# Files I want to read
dir(input_dir, pattern = "xlsx")

# Path to the folder I want to write .csv files to
output_dir <- tempdir()

# Do the work
xl_sheets_to_csv(input_dir, output_dir)

# Confirm
path_file(dir_ls(output_dir, regexp = "new_stem.*csv$"))
}
\author{
Mauro Lepore and Jessica Shue.
}
