% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlff_to_output.R
\name{xlff_to_output}
\alias{xlff_to_output}
\alias{xlff_to_csv}
\alias{xlff_to_xl}
\alias{xlff_to_dfs}
\title{Convert excel-FastField to .csv, .xlsx and a list of dataframes.}
\usage{
xlff_to_csv(input_dir, output_dir = "./", first_census = FALSE)

xlff_to_xl(input_dir, output_dir = "./", first_census = FALSE)

xlff_to_dfs(input_dir, first_census = FALSE)
}
\arguments{
\item{input_dir}{String giving the directory containing the excel workbooks
to read from.}

\item{output_dir}{String giving the directory where to write .csv files to.}

\item{first_census}{Use \code{TRUE} if this is your first census. Use \code{FALSE}
(default) if this is not your first census but a recensus.}
}
\value{
\code{xlff_to_csv()} and \code{xlff_to_xl()} write a .csv or excel (.xlsx) file
per workbook -- combining all spreadsheets. \code{xlff_to_dfs} outputs a list
where each dataframes combines all spreadsheeets of a workbook.
}
\description{
These functions combine spreadsheets from excel workbooks into common data
structures. \code{xlff_to_csv()} and \code{xlff_to_xl()} write a .csv or
excel (.xlsx) file per workbook -- combining all spreadsheets.
\code{xlff_to_dfs} outputs a list of dataframes where each dataframes combines all
spreadsheeets of a workbook.
}
\details{
This is a rigid function with a very specific goal: To process data from a
specific sampling software -- FastField. Specifically, this is what this
function does:
\itemize{
\item Reads each spreadsheet from each workbook and map it to a dataframe.
\item Lowercases and links the names of each dataframe.
\item Adds any missing \strong{key-sheets}:
\itemize{
\item For first census: (1) "root", (2) "multi_stems", (3) "secondary_stems",
and (4) "single_stems".
\item For recensus: (1) "root", (2) "original_stems", (3)
"new_secondary_stems", and (4) "recruits"
}
\item Dates the data by \code{submission_id} (\code{date} comes from the spreadsheet
\code{root}).
\item Lowercases and links the names of each dataframe-variable.
\item Drops fake stems.
\item Output a common data structure of your choice.
}
}
\section{Acknowledgment}{

\itemize{
\item Sabrina Russo helped to make these functions useful with first censuses.
\item David Orwig helped to fix a debug.
}
}

\examples{
library(fs)
library(readr)
library(readxl)

# NOT A FIRST CENSUS
# Path to the folder I want to read excel files from
input_dir <- dirname(tool_example("two_files/new_stem_1.xlsx"))
input_dir

# Files I want to read
dir(input_dir, pattern = "xlsx")

# Path to the folder I want to write .csv files to
output_dir <- tempdir()

# Output a csv file
xlff_to_csv(input_dir, output_dir)

# Confirm
path_file(dir_ls(output_dir, regexp = "new_stem.*csv$"))

# Also possible to output excel and a list of dataframe. See next section.

# FIRST CENSUS
input_dir <- dirname(tool_example("first_census/census.xlsx"))
# As a reminder you'll get a warning of missing sheets
# Output list of dataframes (one per input workbook -- here only one)
dfs <- xlff_to_dfs(input_dir, first_census = TRUE)
str(dfs, give.attr = FALSE)

# Output excel
xlff_to_xl(input_dir, output_dir, first_census = TRUE)
# Read back
filename <- path(output_dir, "census.xlsx")
out <- read_excel(filename)
str(out, give.attr = FALSE)
}
\author{
Mauro Lepore and Jessica Shue.
}
