% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsheets_list.R
\name{xlsheets_list}
\alias{xlsheets_list}
\title{Import mapping each spreadsheet of an excel file to a dataframe in a list.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
xlsheets_list(path)
}
\arguments{
\item{path}{A path to a single excel file.}
}
\value{
A list of dataframes.
}
\description{
A useful complement of this function is \code{\link[=list_csv]{list_csv()}}.
}
\examples{
xlsheets_list(tool_example("multiple_sheets.xlsx"))
}
\seealso{
\code{\link[=list_csv]{list_csv()}}.

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{list_csv}}

Other general functions to import data: \code{\link{dir_list}},
  \code{\link{read_with}}
}
\concept{functions to handle multiple spreadsheets of an excel workbook.}
\concept{general functions to import data}
