% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_status_tree.R
\name{add_status_tree}
\alias{add_status_tree}
\title{Define the status of a tree based on the status of its stems.}
\usage{
add_status_tree(x, status_a = "A", status_d = "D")
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{status_a, status_d}{Sting to match alive and dead stems; it corresponds
to the values of the variable \code{status} (in census tables) or \code{Status} (with
capital "S" in ViewFull tables).}
}
\value{
The input data set with the additional variable \code{status_tree}.
}
\description{
Define the status of a tree based on the status of its stems.
}
\examples{
x <- tibble::tribble(
  ~CensusID, ~Tag, ~Status,
          1,    1, "alive",
          1,    1,  "dead",
          1,    2,  "dead",
          1,    2,  "dead",
          2,    1, "alive",
          2,    1, "alive",
          2,    2, "alive",
          2,    2,  "dead"
)
add_status_tree(x)
}
\seealso{
Other functions to add columns to dataframes.: \code{\link{add_var}}
}
