% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{add_var}
\alias{add_var}
\alias{add_lxly}
\alias{add_index}
\alias{add_rowcol}
\alias{add_hectindex}
\alias{add_quad}
\title{Add columns to a dataframe.}
\usage{
add_var(x, var, gridsize = 20, plotdim = NULL)

add_lxly(x, gridsize = 20, plotdim = NULL)

add_index(x, gridsize = 20, plotdim = NULL)

add_rowcol(x, gridsize = 20, plotdim = NULL)

add_hectindex(x, gridsize = 20, plotdim = NULL)

add_quad(x, gridsize = 20, plotdim = NULL, start = 1, width = 2)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{var}{A character string; either "lxly", "rowcol", "index", hectindex.}

\item{gridsize}{The gridsize of the census plot (commonly 20 m).}

\item{plotdim}{The global dimensions of the census plot (i.e. the
maximum possible values of `gx` and `gy`).}

\item{start}{\code{1} or \code{0}, indicating how to label the first plot-column.}

\item{width}{Number; width to padd the labels of plot-columns and -rows
(passed to \code{\link[stringr:str_pad]{stringr::str_pad()}}).}
}
\value{
A modified version of the dataframe \code{x} with the additional
variable(s) \code{var}.
}
\description{
These functions wrap legacy code from the \href{http://ctfs.si.edu/Public/CTFSRPackage/}{CTFS RPackage}.
}
\examples{
x <- tibble::tibble(gx = c(0, 50, 999.9, 1000), gy = gx/2)

# Each `add_var(x, var = "*")` has a shortcut: `add_*()`
add_var(x, var = "lxly")
# same
add_lxly(x)

add_index(x)

add_rowcol(x)

add_hectindex(x)

add_quad(x)
add_quad(x, start = 0)

# `width` gives the nuber of digits to pad the label of plot-rows and
# plot-columns, e.g. 3 pads plot-rows with three zeros and plot-columns with an
# extra trhree zeros, resulting in a total of 6 zeros.
add_quad(x, start = 0, width = 3)
}
\seealso{
\code{\link[stringr:str_pad]{stringr::str_pad()}}.
}
