% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allquadratslopes.R
\name{allquadratslopes}
\alias{allquadratslopes}
\title{Internal.}
\usage{
allquadratslopes(elev, gridsize = 20, plotdim = c(1000, 500),
  edgecorrect = TRUE)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcslope}},
  \code{\link{findborderquads}}, \code{\link{quadslope}}
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
