% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Hierarchical clustering of topography metrics.}
\usage{
cluster(.data, n)
}
\arguments{
\item{.data}{A dataframe of subclass fgeo_topography.}

\item{n}{Integer. Number of cluster-groups to construct (passed to the
argument \code{k} to \code{\link[stats:cutree]{stats::cutree()}}).}
}
\description{
This functoin adds the column \code{cluster} to objects of class fgeo_topography.
Using the \strong{stats} package, \code{cluster()} calculates a hierarchical clustering
of the columns \code{meanelev}, \code{slope} and \code{convexity}. First, it calculates a
dissimilarities among those columns (with \code{\link[stats:dist]{stats::dist()}} and all its
defaults), then it calculates a tree (with \code{\link[stats:hclust]{stats::hclust()}} and all its
defaults), and finally cuts the tree in \code{n} groups (with \code{\link[stats:cutree]{stats::cutree()}}).
}
\examples{
elev_ls <- fgeo.data::luquillo_elevation
topo <- fgeo_topography(elev_ls, gridsize = 20)
cluster(topo, n = 4)
}
