% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{dir_list}
\alias{dir_list}
\alias{rdata_list}
\alias{rds_list}
\alias{csv_list}
\alias{tsv_list}
\alias{xl_list}
\alias{xlbooks_list}
\title{Import files (.csv, .tsv, excel sheets/workbooks, .Rdata, .rds) into a list.}
\usage{
rdata_list(path_dir, ...)

rds_list(path_dir, ...)

csv_list(path_dir, ...)

tsv_list(path_dir, ...)

xl_list(path_dir, ...)

xlbooks_list(path_dir, ...)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function.}
}
\value{
A list of dataframes.
}
\description{
Import files (.csv, .tsv, excel sheets/workbooks, .Rdata, .rds) into a list.
}
\examples{
path_rdata <- tool_example("rdata")
path_rdata
dir(path_rdata)
rdata_list(path_rdata)

path_rds <- tool_example("rds")
path_rds
dir(path_rds)
rds_list(path_rds)

path_csv <- tool_example("csv")
path_csv
dir(path_csv)
csv_list(path_csv)

path_xl <- tool_example("xl")
path_xl
dir(path_xl)
xl_list(path_xl)

path_books <- tool_example("multiple_workbooks")
dir(path_books)
xlbooks_list(path_books)
}
\seealso{
Other general functions to import data: \code{\link{read_with}},
  \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
