% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{dir_list}
\alias{dir_list}
\alias{rdata_list}
\alias{rda_list}
\alias{rds_list}
\alias{csv_list}
\alias{delim_list}
\alias{tsv_list}
\title{Import multiple files (.csv, excel, .Rdata, ...) from a directory into a list.}
\usage{
rdata_list(path_dir, ...)

rda_list(path_dir, ...)

rds_list(path_dir, ...)

csv_list(path_dir, ...)

delim_list(path_dir, ...)

tsv_list(path_dir, ...)
}
\arguments{
\item{path_dir}{String; the path to a directory containing the files to read
(all must be of appropriate format; see examples).}

\item{...}{Arguments passed to the reader function:
\itemize{
\item \code{rdata_list()} and \code{rda_list()} read with \code{get(load(x))} (\code{...} not unused).
\item \code{rds_list()} reads with \code{\link[readr:read_rds]{readr::read_rds()}}.
\item \code{csv_list()} reads with \code{\link[readr:read_csv]{readr::read_csv()}}.
\item \code{delim_list()} reads with \code{\link[readr:read_delim]{readr::read_delim()}}.
\item \code{tsv_list()} reads with \code{\link[readr:read_tsv]{readr::read_tsv()}}.
}}
}
\value{
A list of dataframes.
}
\description{
These functions read from a specific directory where every file has its own extension
as indicated by each function's name. Notice that function names have the format
input_output, i.e. file-extension_list. If none of these functions do what
you want, create your own with \code{\link[=read_with]{read_with()}}.
}
\examples{
tool_example()

dir(tool_example("rdata"))
rdata_list(tool_example("rdata"))

dir(tool_example("rds"))
rds_list(tool_example("rds"))

dir(tool_example("csv"))
csv_list(tool_example("csv"))

dir(tool_example("tsv"))
tsv_list(tool_example("tsv"))

# Weird: Tab separated columns in a file with .csv extension
dir(tool_example("weird"))

# Extension is .csv, but this is not what you want
csv_list(tool_example("weird"))

# Use this instead
delim_list(tool_example("weird"), delim = "\\t")

}
\seealso{
Other general functions to import data: \code{\link{read_with}}
}
\concept{general functions to import data}
