% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.tool-package.R
\docType{package}
\name{fgeo.tool-package}
\alias{fgeo.tool}
\alias{fgeo.tool-package}
\title{fgeo.tool: Import and Manipulate ForestGEO Data}
\description{
Import and manipulate ForestGEO data
    (<http://www.forestgeo.si.edu/>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.tool}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.tool/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item  Richard  Condit \email{richardcondit@gmail.com}
  \item Suzanne Lao \email{laoz@si.edu}
}

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
