% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_elevation.R
\name{fgeo_elevation}
\alias{fgeo_elevation}
\title{Construct elevation data.}
\usage{
fgeo_elevation(x)
}
\arguments{
\item{x}{Either a dataframe or a dataframe stored in the element \code{col} of a
list.}
}
\value{
A dataframe with names \code{x/gx}, \code{y/gy} and \code{elev}.
}
\description{
This function constructs an object of class fgeo_elevation. It standardizes
the structure of elevation data to always output a dataframe with names \code{gx},
\code{gy} and \code{elev} -- or it dies trying.
}
\section{Input}{

The main input can be either the elevation list that ForestGEO delivers, or
the element \code{col} of such list -- which is a dataframe containing the
elevation data. Notice that the required arguments vary according to the main
input (the elevation list or the elevation dataframe). Whatever the input,
the dataframe containing the elevation data must have columns \code{gx} and \code{gy}
or \code{x} and \code{y}.
}

\section{Acknowledgments}{

This function was inspired by David Kenfack.
}

\examples{
# Dataframe input
elev_df <- bciex::bci_elevation
str(elev_df)

elev <- fgeo_elevation(elev_df)
str(elev)

# List input
elev_list <- list(col = elev_df, other = "stuff")
str(elev_list)

elev <- fgeo_elevation(elev_list)
str(elev) 
}
