% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_habitat.R
\name{fgeo_habitat}
\alias{fgeo_habitat}
\alias{fgeo_habitat.list}
\alias{fgeo_habitat.data.frame}
\title{Construct habitat data.}
\usage{
fgeo_habitat(elevation, gridsize, n, ...)

\method{fgeo_habitat}{list}(elevation, gridsize, n, ...)

\method{fgeo_habitat}{data.frame}(elevation, gridsize, n, xdim, ydim, ...)
}
\arguments{
\item{elevation}{\itemize{
\item A list with at least three elements: \code{col} containing
elevation data; and \code{xdim} and \code{ydim} giving plot dimensions; OR
\itemize{
\item A dataframe containing elevation data, in which
case the parameters \code{xdim} and \code{ydim} must be provided. It may be the
element \code{col} of a ForestGEO elevation-list or an object of class
fgeo_elevation (see \code{\link[=fgeo_elevation]{fgeo_elevation()}}).
}
}}

\item{gridsize}{Number to round \code{x} and \code{y} by. Commonly, \code{gridsize = 20}.}

\item{n}{Number of elevation groups (habitats) to cut elevation data by.
Commonly, \code{n = 4}.}

\item{...}{Other arguments passed to methods.}

\item{xdim, ydim}{\code{x} and \code{y} dimensions of the plot.}
}
\value{
A dataframe of habitat data, with columns \code{gx} and \code{gy}, rounded
with accuracy determined by \code{gridsize}, and column \code{habitats}, with as
many distinct values as determined by the argument \code{n}.
}
\description{
This function constructs an object of class fgeo_habitat. It takes either the
elevation list that ForestGEO delivers, or its element \code{col} -- which
contains the elevation data. Notice that the required arguments to
\code{fgeo_habitat()} vary according to the main input (the elevation list or the
elevation dataframe).
}
\examples{
# Input: Object of class list
elev_list <- fgeo.data::luquillo_elevation
hab1 <- fgeo_habitat(elev_list, gridsize = 20, n = 4)
str(hab1)

# Input: Object of class dataframe
elev_df <- fgeo.data::luquillo_elevation$col
hab2 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab2)

# Input: Object of class fgeo_elev
fgeo_elev <- fgeo_elevation(fgeo.data::luquillo_elevation)
hab3 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab3)

identical(hab1, hab2)
identical(hab2, hab3)
hab1
}
