% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_files_to_df.R
\name{files_to_df}
\alias{files_to_df}
\alias{csv_df}
\alias{xl_df}
\title{Import reducing multiple .csv or flat .xlsx files into a single dataframe.}
\usage{
csv_df(dir, ...)

xl_df(dir, ...)
}
\arguments{
\item{dir}{String giving the directory containing the .csv or excel files to
import.}

\item{...}{Arguments passed to \code{\link[readr:read_csv]{readr::read_csv()}} (\code{csv_*()}) or
\code{\link[readxl:read_excel]{readxl::read_excel()}} (\code{xl_*()}).}
}
\value{
A single dataframe.
}
\description{
Read all .csv (\code{csv_*()}), or excel files (\code{xl_*()}) from a directory into a
single dataframe.

Excel files are treated as if they were flat: This function can read only a
single sheet (specified via the argument \code{sheet} passed to
\code{\link[readxl:read_excel]{readxl::read_excel()}} via \code{...}). For reading multiple all sheets in a
single file see \code{\link[=xlsheets_list]{xlsheets_list()}}.
}
\section{Acknowledgment}{

Thanks to Jessica Shue for inspiring this functions.
}

\examples{
xl_df(tool_example("multiple_workbooks"))

csv_df(tool_example("multiple_csv"))
}
\seealso{
\code{\link[readr:read_csv]{readr::read_csv()}}, \code{\link[readxl:read_excel]{readxl::read_excel()}}.

Other general functions to import data: \code{\link{files_to_list}},
  \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
