% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_files_to_df.R
\name{files_to_df}
\alias{files_to_df}
\alias{csv_to_df}
\alias{csv_to_dfs}
\alias{xl_to_df}
\alias{xl_to_dfs}
\title{Read all .csv or flat excel files from a directory into dataframes.}
\usage{
csv_to_df(dir, ...)

csv_to_dfs(dir, ...)

xl_to_df(dir, ...)

xl_to_dfs(dir, ...)
}
\arguments{
\item{dir}{String giving the directory containing the excel workbooks
to read from.}

\item{...}{Arguments passed to \code{\link[readr:read_csv]{readr::read_csv()}} (\code{csv_*()}) or
\code{\link[readxl:read_excel]{readxl::read_excel()}} (\code{xl_*()}).}
}
\value{
The versions ending in \code{_df} output a single dataframe. The versions
ending in \code{_dfs} output a list of dataframes.
}
\description{
Read all .csv (\code{csv_*()}), or excel files (\code{xl_*()}) from a directory into a
single dataframe (\code{*_df()}) or a list of dataframes (\code{*_dfs()}) -- where each
element of the list is named as the source file. Excel files are treated as
if they were flat -- meaning that these \code{xl_*()} functions can read only a
single sheet (specified via the argument \code{sheet} passed to
\code{\link[readxl:read_excel]{readxl::read_excel()}} via \code{...}). For reading multiple all sheets in a single
file see \code{\link[=xlsheets_to_dfs]{xlsheets_to_dfs()}}.
}
\section{Acknowledgment}{

Thanks to Jessica Shue for inspiring this functions.
}

\examples{
xl_to_dfs(tool_example("multiple_workbooks"))

xl_to_df(tool_example("multiple_workbooks"))

# Pass an argument to `read_excel()` via `...`
xl_to_dfs(tool_example("multiple_workbooks"), sheet = 2)

csv_to_dfs(tool_example("multiple_csv"))

csv_to_df(tool_example("multiple_csv"))

# Pass an argument to `read_csv()` via `...`
csv_to_dfs(tool_example("multiple_csv"), n_max = 2)
}
\seealso{
\code{\link[readr:read_csv]{readr::read_csv()}}, \code{\link[readxl:read_excel]{readxl::read_excel()}}.

Other general functions to import/export data: \code{\link{dfs_to_csv}}
}
\concept{general functions to import/export data}
