% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill NA's with a single filler.}
\usage{
fill_na(x, filler = 0)
}
\arguments{
\item{x}{A dataframe.}

\item{filler}{A filler; whatever you want to replace NA's with.}
}
\value{
A modified version of \code{x}.
}
\description{
This is a wrapper of \code{\link[base:replace]{base::replace()}} with a simplified interface and a name
that target the specific job of filling NA's.
}
\section{Acknowledgments}{

This function was inspired by conversations with David Kenfack and by a
\href{https://goo.gl/aMTQck}{blog post by Rick Pack}. Also tanks to Edgar Ruiz
(@edgararuiz) for discussing alternatives
(via \href{https://goo.gl/4uME2U}{this blog post} on RStudio community).
}

\examples{
x <- data.frame(x = c(NA, 1), y = c("a", NA), stringsAsFactors = FALSE)

a_dataframe <- x
fill_na(a_dataframe)
fill_na(a_dataframe, "")
fill_na(a_dataframe, "missing")

a_matrix <- as.matrix(x)
fill_na(a_matrix)

a_vector <- x$x
fill_na(a_vector)

a_list <- list(x, x, x)
lapply(a_list, fill_na)
}
