% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allquadratslopes.R
\name{findborderquads}
\alias{findborderquads}
\title{Internal.}
\usage{
findborderquads(index, dist = 20, gridsize = 20, plotdim = c(1000,
  500))
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{allquadratslopes}},
  \code{\link{calcslope}}, \code{\link{quadslope}}
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
