% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_manipulate_names.R
\name{nms_extract_anycase}
\alias{nms_extract_anycase}
\title{Find a name exactly matching a string but regardless of case.}
\usage{
nms_extract_anycase(x, nm)
}
\arguments{
\item{x}{A named object.}

\item{nm}{A string to match names exactly but regardless of case.}
}
\value{
A string of the name that was found in \code{names(x)}.
}
\description{
Find a name exactly matching a string but regardless of case.
}
\examples{
v <- c(a = 1, B = 1)
nms_extract_anycase(v, "b")

dfm <- data.frame(a = 1, B = 1)
nms_extract_anycase(dfm, "b")
}
\seealso{
Other functions for developers.: \code{\link{detect_duplicated_var}},
  \code{\link{find_data_of_class}},
  \code{\link{flag_duplicated_var}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}},
  \code{\link{type_ensure}}

Other functions dealing with names.: \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms_tidy}},
  \code{\link{nms_try_rename}}, \code{\link{nms}}
}
\concept{functions dealing with names.}
\concept{functions for developers.}
