% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{nms_minus_lower_nms}
\alias{nms_minus_lower_nms}
\title{Comparing two dataframes, How many names differ only in case?}
\usage{
nms_minus_lower_nms(table1, table2)
}
\arguments{
\item{table1}{A dataframe}

\item{table2}{A dataframe}
}
\value{
An number indicating how many names are different only in their case.
}
\description{
Comparing two dataframes, How many names differ only in case?
}
\examples{
\dontrun{
vft <- yosemite::ViewFullTable_yosemite
stem <- yosemite::yosemite_s1_lao
tree <- yosemite::yosemite_f1_lao
nms_minus_lower_nms(stem, vft)
nms_minus_lower_nms(vft, stem)
nms_minus_lower_nms(stem, tree)
}
}
\seealso{
Other functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}, \code{\link{exists_in_pkg}},
  \code{\link{fgeo_package_deps}},
  \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}}

Other functions to manipulate names.: \code{\link{nms_extract_anycase}},
  \code{\link{nms_lowercase}},
  \code{\link{nms_restore_newvar}}, \code{\link{nms}}
}
