% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick rows from ForestGEO censuses in a list.}
\usage{
pick(.data, ..., key = NULL)
}
\arguments{
\item{.data}{A ForestGEO dataset of class 'censuses_lst' (a list).}

\item{...}{Other arguments passed to methods.}

\item{key}{Key dataframe to pick rows from and recycle in all other
censuses.}
}
\description{
This function allows you to pick rows from a key census (dataframe) in a list
and pick the exact same rows in all other non-key censuses.
}
\examples{
censuses_lst <- as_censuses(list(
  c1 = tibble(dbh = 1:2),
  c2 = tibble(dbh = 8:9)
))
censuses_lst
class(censuses_lst)

pick(censuses_lst, dbh == 1)
pick(censuses_lst, dbh >= 2)
pick(censuses_lst, dbh <= 8 , key = "c2")

path <- tool_example("rdata")
dir(path)
censuses <- as_censuses(rdata_list(path))
pick(censuses, dbh > 30)
}
\seealso{
Other functions to pick or drop rows of a ForestGEO dataframe: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}}, \code{\link{pick_drop}},
  \code{\link{pick_main_stem}}, \code{\link{pick_plotname}}
}
\concept{functions to pick or drop rows of a ForestGEO dataframe}
