% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_in_sync.R
\name{pick}
\alias{pick}
\title{Pick a list of datframes or a nested dataframe in sync.}
\usage{
pick(.data, ..., key = NULL)
}
\arguments{
\item{.data}{A list or nested dataframe.}

\item{...}{Other arguments passed to methods.}

\item{key}{Key dataframe to pick rows from and use the same row indices to
pick rows from all other dataframes in the list (or list-column).}
}
\description{
This function allows you to pick rows from a key dataframe in a list or
list-column and pick the exact same rows in all other non-key dataframes
of the list (or list-column).
}
\examples{
lst <- list(
  c1 = tibble::tibble(dbh = 1:2),
  c2 = tibble::tibble(dbh = 8:9)
)

pick(lst, dbh == 1)
pick(lst, dbh >= 2)
pick(lst, dbh <= 8 , key = 2)

dfm <- tibble::tribble(
  ~dbh, ~census,
    1L,       1,
    2L,       1,
    8L,       2,
    9L,       2
)

nested_df <- dfm \%>\% dplyr::group_by(census) \%>\% nest()

out <- pick(nested_df, dbh == 1)
out
# Same but ackward
out$data

out <- pick(nested_df, dbh >= 1)
out
# Same but ackward
out$data


censuses <- rdata_df(tool_example("rdata"), .id = "census") \%>\% 
  group_by(census) \%>\% 
  nest()
censuses

pick(censuses, dbh > 30, key = 1)
}
