% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick rows from ForestGEO censuses in a list.}
\usage{
pick(.data, ..., key = NULL)
}
\arguments{
\item{.data}{A ForestGEO dataset of class 'censuses_lst' (a list).}

\item{...}{Other arguments passed to methods.}

\item{key}{Key dataframe to pick rows from and recycle in all other
censuses.}
}
\description{
This function allows you to pick rows from a key census (dataframe) in a list
and pick the exact same rows in all other non-key censuses.
}
\examples{
censuses <- as_censuses(
  list(
    c1 = tibble(dbh = 1:2),
    c2 = tibble(dbh = 8:9)
  )
)
class(censuses)

censuses

pick(censuses, dbh == 1)

pick(censuses, dbh >= 2)

pick(censuses, dbh >= 2, dbh < 9)

# `key` defaults to the first element of the list
pick(censuses, dbh <= 8)

# Different key
pick(censuses, dbh <= 8 , key = "c2")

# `pick()` is a common step after reading multiple censuses into a list
rdata_files <- tool_example("rdata")
dir(rdata_files)

censuses <- as_censuses(rdata_list(rdata_files))
pick(censuses, dbh >= 100)

# Same as a pipeline
rdata_files \%>\% 
  rdata_list() \%>\% 
  as_censuses() \%>\% 
  pick(dbh >= 100)

}
\seealso{
Other functions to pick or drop rows of a ForestGEO dataframe: \code{\link{pick_drop}},
  \code{\link{pick_main_stem}}, \code{\link{pick_plotname}}
}
\concept{functions to pick or drop rows of a ForestGEO dataframe}
