% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports-reexports-fgeo-base.R
\name{pick_drop}
\alias{pick_drop}
\alias{pick_dbh_min}
\alias{pick_dbh_max}
\alias{pick_dbh_under}
\alias{pick_dbh_over}
\alias{pick_status}
\alias{drop_status}
\title{Pick and drop rows of a ForestGEO ViewFullTable or census table.}
\usage{
pick_dbh_min(dfm, x, na.rm = FALSE)

pick_dbh_max(dfm, x, na.rm = FALSE)

pick_dbh_under(dfm, x, na.rm = FALSE)

pick_dbh_over(dfm, x, na.rm = FALSE)

pick_status(dfm, x, na.rm = FALSE)

drop_status(dfm, x, na.rm = FALSE)
}
\arguments{
\item{dfm}{Dataframe; A ForestGEO stem- or tree-table.}

\item{x}{An atomic vector; a single value against to compare each value of
the variable encoded in the function name.}

\item{na.rm}{Set to \code{TRUE} if you want to remove missing values from the
variable encoded in the function name.}
}
\value{
Dataframe rows with matching conditions.
}
\description{
These functions provide an expressive and convenient way to filter
ForestGEO's data. They let you remove missing values (with \code{na.rm = TRUE})
but conservatively default to keeping NAs. This behavior is similar to base
subsetting and unlike \code{dplyr::filter()}; this conservative default matters
because the difference is important between missing trees and dead trees --
and you may want to include missing trees in your analysis.
}
\examples{
cns <- data.frame(
  dbh = c(0, 50, 100, 150, NA, NA, NA),
  status = c(rep("A", 4), "M", "D", NA)
  # stringsAsFactors = FALSE
)
cns

# <=
pick_dbh_max(cns, 100)
pick_dbh_max(cns, 100, na.rm = TRUE)
# >=
pick_dbh_min(cns, 100)
pick_dbh_min(cns, 100, na.rm = TRUE)
# <
pick_dbh_under(cns, 100)
pick_dbh_under(cns, 100, na.rm = TRUE)
# >
pick_dbh_over(cns, 100)
pick_dbh_over(cns, 100, na.rm = TRUE)
# Same, but `subset()` does not let you keep NAs.
subset(cns, dbh > 100)

# ==
pick_status(cns, "A")
pick_status(cns, "A", na.rm = TRUE)
# !=
drop_status(cns, "D")
drop_status(cns, "D", na.rm = TRUE)

# Compose
pick_dbh_over(drop_status(cns, "D", na.rm = TRUE), 100)
}
\seealso{
Other functions to pick or drop rows of a ForestGEO dataframe: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}},
  \code{\link{pick_main_stem}}, \code{\link{pick}}
}
\concept{functions to pick or drop rows of a ForestGEO dataframe}
