% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_largest_hom_dbh.R
\name{pick_largest_hom_dbh}
\alias{pick_largest_hom_dbh}
\title{Pick the main stem of each tree in each census.}
\usage{
pick_largest_hom_dbh(.x)
}
\arguments{
\item{.x}{A ForestGEO-like dataframe, census or ViewFullTable.}
}
\value{
A dataframe with one row per per treeid per censusid and a single
plotname.
}
\description{
This function picks the main stem of each tree in each census.
}
\details{
This function picks the main stem of each tree in each census. It
collapses data of multi-stem trees by picking a single stem per \code{treeid} per
\code{censusid}: Within this groups it picks the stem at the top of a list sorted
first by descending order of \code{hom}, and then by descending order of \code{dbh} --
this corrects the effect of buttresses and picks the main stem. It
intentionally rejects data with multiple plots, and ignores groups of grouped
data.
}
\section{Warning}{

Currently this function is considerably slow.
}

\examples{
library(tibble)

# Trees with buttresses may have more than one measurements per stem.
census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,  
  "sp1",     "1",   "1.2",   130,   55,         1,  
  "sp2",     "2",   "2.1",   130,    5,         1   # main stem
)

# Piks largest hom first (to correct effect of batreesses) then largest dbh
pick_largest_hom_dbh(census)
}
\seealso{
Other functions to pick or drop rows of a dataframe.: \code{\link{drop_twice_dead}},
  \code{\link{filter_status}}, \code{\link{pick_recensus}},
  \code{\link{pick_top}}
}
\concept{functions to pick or drop rows of a dataframe.}
