% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_woods.R
\name{pick_woods}
\alias{pick_woods}
\title{Pick woods after collapsing multiple stems by treeid and censusid.}
\usage{
pick_woods(.data, ...)
}
\description{
Pick woods after collapsing multiple stems by treeid and censusid.
}
