% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_woods.R
\name{pick_woods}
\alias{pick_woods}
\alias{pick_trees}
\alias{pick_saplings}
\alias{pick_saplings_and_trees}
\title{Pick main stems within a specific dbh range.}
\usage{
pick_woods(.data, ...)

pick_trees(.data)

pick_saplings(.data)

pick_saplings_and_trees(.data)
}
\arguments{
\item{.data}{A ForestGEO-like dataframe, either a census or ViewFullTable.}

\item{...}{Expressions passed to \code{dplyr::filter()}, E.g. the
expression \code{dbh >= 100} picks \code{dbh} values of 100 mm and above.}
}
\value{
A dataframe. See description.
}
\description{
These functions extend \code{pick_main_stem()}. Working with simple or
grouped dataframes (via \code{dplyr::group_by()}) these functions let you pick
main stems within a given \code{dbh} range:
\itemize{
\item \code{pick_woods()} is a general function and lets you choose the \code{dbh} range. The
other functions are shortcuts with pre-defined \code{dbh} ranges:
\itemize{
\item \code{pick_saplings_and_trees()} picks stems of 10 mm dbh and above.
\item \code{pick_saplings()} picks stems between 10 mm dbh inclusive and 100 mm dbh
exclusive.
\item \code{pick_trees()} picks stems of 100 mm dbh and above.
}
}
}
\examples{
library(tibble)

census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1   # buttress
)

# Trees with buttresses may have more than one measurements per stem.
# Piks largest hom first (to correct effect of batreesses) then largest dbh
pick_main_stem(census)

# Also works with multiple censuses
censuses <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,  # buttress
  "sp1",     "1",   "1.1",   140,   50,         2,  # main stem
  "sp1",     "1",   "1.1",   130,   70,         2   # buttress
)

pick_main_stem(censuses)

# pick_woods() and friends internally pick the main stem, then pick by dbh
pick_woods(censuses)

pick_woods(censuses, dbh > 40)

# These functions are convenient shortcuts
census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,  # buttress
  "sp1",     "2",   "2.1",   130,   100,        1,
  "sp1",     "3",   "2.1",   130,   110,        1,
)

pick_saplings(census)
pick_trees(census)
pick_saplings_and_trees(census)
}
\concept{functions for fgeo census and vft.}
