% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fgeo.R
\name{read_fgeo}
\alias{read_fgeo}
\alias{read_vft}
\alias{read_taxa}
\title{Read ViewFullTable and ViewTaxonomy files delivered by ForestGEO's database.}
\usage{
read_vft(file, delim = "\\t", na = c("", "NA", "NULL"), ...)

read_taxa(file, delim = "\\t", na = c("", "NA", "NULL"), ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path).}

\item{delim}{Single character used to separate fields within a record.}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{...}{Other arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A dataframe (or tibble).
}
\description{
These functions pre-fill arguments of \code{\link[readr:read_delim]{readr::read_delim()}} to avoid the most
common causes of problems when reading ForestGEO data as delivered by the
database. In most cases you should only need to provide the argument \code{file}.
}
\examples{

}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}}, \code{\link[=type_vft]{type_vft()}},  \code{\link[=type_taxa]{type_taxa()}}.

Other functions to read text files delivered by ForestgGEO's database.: \code{\link{type_fgeo}}
}
