% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fgeo.R
\name{read_fgeo}
\alias{read_fgeo}
\alias{read_vft}
\alias{read_taxa}
\title{Import ViewFullTable and ViewTaxonomy from .tsv or .csv files.}
\usage{
read_vft(file, delim = NULL, na = c("", "NA", "NULL"), ...)

read_taxa(file, delim = NULL, na = c("", "NA", "NULL"), ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[=clipboard]{clipboard()}} will read from the system clipboard.}

\item{delim}{Single character used to separate fields within a record. The
default (\code{delim = NULL}) is to guess between comma or tab (\code{","} or \code{"\t"}).}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{...}{Other arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A dataframe (or tibble).
}
\description{
These functions help you read text files containing ViewFullTable and
ViewTaxonomy tables. They avoid common problems related to column separators,
missing values, column names, and column types. Although you can customize
them, in most cases you should only need to provide the argument \code{file}.
}
\section{Acknowledgments}{

Thanks to Shameema Jafferjee Esufali for inspiring the feature that
automatically detects \code{delim} (issue 65).
}

\examples{
\dontrun{
# Needs internet connection.
read_vft("http://bit.ly/fgeo-data-luquillo-vft-random")
read_taxa("http://bit.ly/fgeo-data-luquillo-taxa")
}
}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}}, \code{\link[=type_vft]{type_vft()}}, \code{\link[=type_taxa]{type_taxa()}}.

Other functions to read text files delivered by ForestgGEO's database.: \code{\link{type_fgeo}}

Other functions to import ForestGEO data: \code{\link{read_censuses}}
}
\concept{functions to import ForestGEO data}
\concept{functions to import a single file of ForestGEO data}
\concept{functions to read text files delivered by ForestgGEO's database.}
