% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{read_with}
\alias{read_with}
\title{Import multiple files into a list using any given reading function.}
\usage{
read_with(.f, regexp = NULL)
}
\arguments{
\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
}
\value{
A modified version of the input function, able to read all files
from a directory (provided they all are of the suitable format).
}
\description{
Import multiple files into a list using any given reading function.
}
\examples{
path_rds <- tool_example("rds")
path_rds
dir(path_rds)

rds_list <- read_with(readr::read_rds)
rds_list(path_rds)

# Same
read_with(readr::read_rds)(path_rds)

# Read specific files by matching file names via `regexp`
dir(path_rds)
read_with(readr::read_rds, regexp = "6")(path_rds)
read_with(readr::read_rds, regexp = "5[.]rds")(path_rds)
read_with(readr::read_rds, regexp = "tree.*[.]rds")(path_rds)

# Read excel files --------------------------------------------------------
\dontrun{
if (!requireNamespace("readxl"))
  stop("Please install readxl with `install.packges('readxl')`")
path_xl <- tool_example("xl")
dir(path_xl)
read_with(readxl::read_excel)(path_xl)
}

# Read mixed files --------------------------------------------------------
\dontrun{
if (!requireNamespace("rio"))
  stop("Please install rio with `install.packges('rio')`")
if (!requireNamespace("readxl"))
  stop("Please install readxl with `install.packges('readxl')`")

path_mixed_files <- tool_example("mixed_files")
dir(path_mixed_files)
read_with(rio::import)(path_mixed_files)
}
}
\seealso{
Other general functions to import data: \code{\link{dir_list}},
  \code{\link{rdata_df}}
}
\concept{general functions to import data}
