% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{read_with}
\alias{read_with}
\title{Import multiple files into a list using any given reading function.}
\usage{
read_with(.f, regexp = NULL)
}
\arguments{
\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
}
\value{
A modified version of the input function, able to read all files
from a directory (provided they all are of the suitable format).
}
\description{
Import multiple files into a list using any given reading function.
}
\examples{
path_rds <- tool_example("rds")
path_rds
dir(path_rds)

rds_list <- read_with(readr::read_rds)
rds_list(path_rds)

# Same
read_with(readr::read_rds)(path_rds)

\dontrun{
if (!requireNamespace("rio"))
  stop("Please install rio with install.packges('rio') to run this example")

guess_list <- read_with(rio::import)
path_mixed_ext <- tool_example("mixed_files")
dir(path_mixed_ext)
guess_list(path_mixed_ext)
}
}
\seealso{
Other general functions to import data: \code{\link{dir_list}},
  \code{\link{xlsheets_list}}
}
\concept{general functions to import data}
\keyword{internal}
