% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{read_with}
\alias{read_with}
\title{Import multiple files into a list using any given reading function.}
\usage{
read_with(.f, regexp = NULL)
}
\arguments{
\item{.f}{A function able to read the desired file format.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
}
\value{
A modified version of the input function, able to read all files
from a directory (provided they all are of the suitable format).
}
\description{
Import multiple files into a list using any given reading function.
}
\examples{
library(fgeo.x)

rds_file <- path_example("rds")
dir(rds_file)

rds_list <- read_with(readr::read_rds)
rds_list(rds_file)

# Same
read_with(readr::read_rds)(rds_file)

# Read specific files by matching file names via `regexp`
dir(rds_file)
read_with(readr::read_rds, regexp = "6")(rds_file)
read_with(readr::read_rds, regexp = "5[.]rds")(rds_file)
read_with(readr::read_rds, regexp = "tree.*[.]rds")(rds_file)

# Read excel files --------------------------------------------------------
excel_file <- path_example("xl")
dir(excel_file)
\dontrun{
readxl_is_installed <- requireNamespace("readxl", quietly = TRUE)
if (readxl_is_installed) {
  library(readxl)
  
  read_with(readxl::read_excel)(excel_file)
}
}

# Read mixed files --------------------------------------------------------
mixed_files <- path_example("mixed_files")
dir(mixed_files)
\dontrun{
readxl_is_installed <- requireNamespace("readxl", quietly = TRUE)
rio_is_installed <- requireNamespace("rio", quietly = TRUE)
if (rio_is_installed && readxl_is_installed) {
  library(rio)
  library(readxl)
  
  read_with(rio::import)(mixed_files)
}
}
}
\seealso{
Other general functions to import data: \code{\link{dir_list}}
}
\concept{general functions to import data}
