% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-imports-reeports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{filter}
\alias{mutate}
\alias{select}
\alias{arrange}
\alias{summarize}
\alias{summarise}
\alias{group_by}
\alias{ungroup}
\alias{count}
\alias{add_count}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\alias{last_col}
\alias{tibble}
\alias{tribble}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{add_count}}}

  \item{rlang}{\code{\link[rlang]{\%||\%}}}

  \item{tibble}{\code{\link[tibble]{tibble}}, \code{\link[tibble]{tribble}}, \code{\link[tibble]{as_tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}}
}}

