% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_all_na.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace all NA's of a dataframe with a filler}
\usage{
replace_all_na(x, filler = 0)
}
\arguments{
\item{x}{A dataframe.}

\item{filler}{A filler; whatever you want to replace NA's with.}
}
\value{
A modified versio of \code{x}.
}
\description{
Replace all NA's of a dataframe with a filler
}
\section{Acknowledgements}{

This function was inspired by David Kenfack.
}

\examples{
x <- data.frame(x = c(NA, 1), y = c("a", NA), stringsAsFactors = FALSE)
replace_all_na(x)
replace_all_na(x, "")
replace_all_na(x, "missing")
}
