% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{standardize_at}
\alias{standardize_at}
\title{Divide (standardize) columns of a dataframe by a constant denominator.}
\usage{
standardize_at(x, .at, denominator)
}
\arguments{
\item{x}{Dataframe.}

\item{.at}{A character vector of names or a numeric vector of
positions. Only those elements corresponding to \code{.at} will be
modified.}

\item{denominator}{A numeric vector of length 1.}
}
\value{
A data.frame.
}
\description{
Divide (standardize) columns of a dataframe by a constant denominator.
}
\examples{
dfm <- data.frame(a = 1:3, b = 11:13)
standardize_at(dfm, "a", denominator = 100)
standardize_at(dfm, names(dfm), denominator = 100)
}
\seealso{
Other general functions to edit data in place: \code{\link{convert_unit}},
  \code{\link{fill_na}}, \code{\link{lookup}},
  \code{\link{to_tidy_names}}
}
\concept{general functions to edit data in place}
