% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_status.R
\name{status_stem}
\alias{status_stem}
\title{Filter a data set by the status of its stems.}
\usage{
status_stem(x, .status)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{.status}{Character vector; Must be one of possible values of the
variable giving the status of the dataframe \code{x}.}
}
\description{
Filter a data set by the status of its stems.
}
