% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_tree.R
\name{status_tree}
\alias{status_tree}
\title{Define the status of a tree based on the status of its stems.}
\usage{
status_tree(x)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}
}
\value{
The input data set with the additional variable \code{status_tree}.
}
\description{
Define the status of a tree based on the status of its stems.
}
\examples{
x <- tibble::tribble(
  ~CensusID, ~Tag, ~Status,
          1,    1, "alive",
          1,    1,  "dead",
          1,    2,  "dead",
          1,    2,  "dead",
          2,    1, "alive",
          2,    1, "alive",
          2,    2, "alive",
          2,    2,  "dead"
)
status_tree(x)
}
