% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.demography_lst}
\alias{to_df.demography_lst}
\alias{to_df.demography_lst_by}
\title{Transform the output of demography functions into a (tibble) dataframe.}
\usage{
\method{to_df}{demography_lst}(.x)

\method{to_df}{demography_lst_by}(.x)
}
\arguments{
\item{.x}{An object of class demography_lst.}

\item{...}{Other arguments passed to \code{\link[=demography_df]{demography_df()}}.}
}
\value{
A (tibble) dataframe.
}
\description{
Restructure  the output of \code{mortality()}, \code{recruitment()}, and \code{growth()} as
a dataframe:
\itemize{
\item \code{to_df.demography_lst()}: Restructures results calculated across the entire
census data.
\item \code{to_df.demography_lst_by()}: Restructures results calculated \code{by} groups.
}
}
\examples{
\dontrun{
library(fgeo.data)
library(fgeo.demography)

censuses <- list(luquillo_tree5_random, luquillo_tree6_random)
to_df(mortality(censuses))

to_df(mortality(censuses, "sp"))
}
}
