% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.tt_lst}
\alias{to_df.tt_lst}
\title{Dataframe objects of class "tt_lst".}
\usage{
\method{to_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_lst.}

\item{...}{Other arguments passed to \code{\link[=to_df]{to_df()}}.}
}
\value{
A dataframe.
}
\description{
This method creates a dataframe from the output of \code{fgeo.habitat::tt_test()}
(which is a list of class "tt_lst").
}
\examples{
if (requireNamespace("fgeo.habitat")) {
  library(fgeo.habitat)
  
  cns <- fgeo.habitat::luquillo_top3_sp
  hab <- fgeo.data::luquillo_habitat
  
  tt_lst <- tt_test(cns, unique(cns$sp), hab)
  to_df(tt_lst)
}
}
\seealso{
\code{\link[=to_df]{to_df()}}.

Other methods for fgeo generics: \code{\link{to_df.demography_lst}},
  \code{\link{to_df.krig_lst}}
}
\concept{methods for fgeo generics}
