% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports-reexports-fgeo-base.R
\name{to_tidy_names}
\alias{to_tidy_names}
\title{Create a names-friendly version of a string (lowercase and with no spaces).}
\usage{
to_tidy_names(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A modified version of \code{x}.
}
\description{
Create a names-friendly version of a string (lowercase and with no spaces).
}
\examples{
messy <- "Hi yOu"

to_tidy_names(messy)

messy_named_string <- c(`Messy Name` = messy)
# Targets strings, not its names (Target names with `fgeo.tool::nms_tidy()`)
to_tidy_names(messy_named_string)

dfm <- data.frame(1)
setNames(dfm, to_tidy_names(messy))

# Makes more sense when operating on strings
setNames(list(1), to_tidy_names(messy))
}
\seealso{
Other general functions to edit data in place: \code{\link{convert_unit}},
  \code{\link{fill_na}}, \code{\link{lookup}},
  \code{\link{standardize_at}}
}
\concept{general functions to edit data in place}
