% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_treeid.R
\name{warn_duplicated_treeid}
\alias{warn_duplicated_treeid}
\alias{detect_duplicated_treeid}
\title{Detect and warn duplicated treeid.}
\usage{
warn_duplicated_treeid(.data)

detect_duplicated_treeid(.data)
}
\arguments{
\item{.data}{A ForestGEO dataframe -- census or ViewFullTable.}
}
\value{
\itemize{
\item \code{warn_duplicated_treeid()}: A warning or invisible \code{.data}.
\item \code{detect_duplicated_treeid()}: \code{TRUE} or \code{FALSE}.
}
}
\description{
Detect and warn duplicated treeid.
}
\examples{
detect_duplicated_treeid(fgeo.data::luquillo_stem6_1ha)
warn_duplicated_treeid(fgeo.data::luquillo_stem6_1ha)

detect_duplicated_treeid(fgeo.data::luquillo_tree6_1ha)
warn_duplicated_treeid(fgeo.data::luquillo_tree6_1ha)

combo56 <- purrr::reduce(
  list(fgeo.data::luquillo_tree5_random, fgeo.data::luquillo_tree5_random),
  rbind
)
# Silent because it cheks for each censusid
detect_duplicated_treeid(combo56)
warn_duplicated_treeid(combo56)
}
