% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_files_to_df.R
\name{xlsheets_to_dfs}
\alias{xlsheets_to_dfs}
\title{Read one excel workbook and map each spreadsheet to a dataframe in a list.}
\source{
Adapted from an article by Jenny Bryan (https://goo.gl/ah8qkX).
}
\usage{
xlsheets_to_dfs(path)
}
\arguments{
\item{path}{A path to an excel file.}
}
\value{
A list of dataframes.
}
\description{
A useful complement of this function is \code{\link[=dfs_to_csv]{dfs_to_csv()}}.
}
\examples{
xlsheets_to_dfs(tool_example("multiple_sheets.xlsx"))
}
\seealso{
\code{\link[=dfs_to_csv]{dfs_to_csv()}}.

Other functions to handle multiple spreadsheets of an excel workbook.: \code{\link{dfs_to_csv}}
}
\concept{functions to handle multiple spreadsheets of an excel workbook.}
